Type Alias esp32::emac_dma::dmain_en::W

pub type W = W<DMAIN_EN_SPEC>;

Register DMAIN_EN writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dmain_tie(&mut self) -> DMAIN_TIE_W<'_, DMAIN_EN_SPEC>

Bit 0 - When this bit is set with Normal Interrupt Summary Enable (Bit[16]) the Transmit Interrupt is enabled. When this bit is reset the Transmit Interrupt is disabled.

pub fn dmain_tse(&mut self) -> DMAIN_TSE_W<'_, DMAIN_EN_SPEC>

Bit 1 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Transmission Stopped Interrupt is enabled. When this bit is reset the Transmission Stopped Interrupt is disabled.

pub fn dmain_tbue(&mut self) -> DMAIN_TBUE_W<'_, DMAIN_EN_SPEC>

Bit 2 - When this bit is set with Normal Interrupt Summary Enable (Bit 16) the Transmit Buffer Unavailable Interrupt is enabled. When this bit is reset the Transmit Buffer Unavailable Interrupt is Disabled.

pub fn dmain_tjte(&mut self) -> DMAIN_TJTE_W<'_, DMAIN_EN_SPEC>

Bit 3 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Transmit Jabber Timeout Interrupt is enabled. When this bit is reset the Transmit Jabber Timeout Interrupt is disabled.

pub fn dmain_oie(&mut self) -> DMAIN_OIE_W<'_, DMAIN_EN_SPEC>

Bit 4 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Receive Overflow Interrupt is enabled. When this bit is reset the Overflow Interrupt is disabled.

pub fn dmain_uie(&mut self) -> DMAIN_UIE_W<'_, DMAIN_EN_SPEC>

Bit 5 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Transmit Underflow Interrupt is enabled. When this bit is reset the Underflow Interrupt is disabled.

pub fn dmain_rie(&mut self) -> DMAIN_RIE_W<'_, DMAIN_EN_SPEC>

Bit 6 - When this bit is set with Normal Interrupt Summary Enable (Bit[16]) the Receive Interrupt is enabled. When this bit is reset the Receive Interrupt is disabled.

pub fn dmain_rbue(&mut self) -> DMAIN_RBUE_W<'_, DMAIN_EN_SPEC>

Bit 7 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Receive Buffer Unavailable Interrupt is enabled. When this bit is reset the Receive Buffer Unavailable Interrupt is disabled.

pub fn dmain_rse(&mut self) -> DMAIN_RSE_W<'_, DMAIN_EN_SPEC>

Bit 8 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Receive Stopped Interrupt is enabled. When this bit is reset the Receive Stopped Interrupt is disabled.

pub fn dmain_rwte(&mut self) -> DMAIN_RWTE_W<'_, DMAIN_EN_SPEC>

Bit 9 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Receive Watchdog Timeout Interrupt is enabled. When this bit is reset the Receive Watchdog Timeout Interrupt is disabled.

pub fn dmain_etie(&mut self) -> DMAIN_ETIE_W<'_, DMAIN_EN_SPEC>

Bit 10 - When this bit is set with an Abnormal Interrupt Summary Enable (Bit[15]) the Early Transmit Interrupt is enabled. When this bit is reset the Early Transmit Interrupt is disabled.

pub fn dmain_fbee(&mut self) -> DMAIN_FBEE_W<'_, DMAIN_EN_SPEC>

Bit 13 - When this bit is set with Abnormal Interrupt Summary Enable (Bit[15]) the Fatal Bus Error Interrupt is enabled. When this bit is reset the Fatal Bus Error Enable Interrupt is disabled.

pub fn dmain_erie(&mut self) -> DMAIN_ERIE_W<'_, DMAIN_EN_SPEC>

Bit 14 - When this bit is set with Normal Interrupt Summary Enable (Bit[16]) the Early Receive Interrupt is enabled. When this bit is reset the Early Receive Interrupt is disabled.

pub fn dmain_aise(&mut self) -> DMAIN_AISE_W<'_, DMAIN_EN_SPEC>

Bit 15 - When this bit is set abnormal interrupt summary is enabled. When this bit is reset the abnormal interrupt summary is disabled. This bit enables the following interrupts in Status Register: Bit[1]: Transmit Process Stopped. Bit[3]: Transmit Jabber Timeout. Bit[4]: Receive Overflow. Bit[5]: Transmit Underflow. Bit[7]: Receive Buffer Unavailable. Bit[8]: Receive Process Stopped. Bit[9]: Receive Watchdog Timeout. Bit[10]: Early Transmit Interrupt. Bit[13]: Fatal Bus Error.

pub fn dmain_nise(&mut self) -> DMAIN_NISE_W<'_, DMAIN_EN_SPEC>

Bit 16 - When this bit is set normal interrupt summary is enabled. When this bit is reset normal interrupt summary is disabled. This bit enables the following interrupts in Status Register: Bit[0]: Transmit Interrupt. Bit[2]: Transmit Buffer Unavailable. Bit[6]: Receive Interrupt. Bit[14]: Early Receive Interrupt.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual