Module esp32::emac_dma::dmaoperation_mode

Receive and Transmit operating modes and command

Structs

DMAOPERATION_MODE_SPEC

Receive and Transmit operating modes and command

Type Aliases

DIS_DROP_TCPIP_ERR_FRAM_R

Field DIS_DROP_TCPIP_ERR_FRAM reader - When this bit is set the MAC does not drop the frames which only have errors detected by the Receive Checksum engine.When this bit is reset all error frames are dropped if the Fwd_Err_Frame bit is reset.

DIS_DROP_TCPIP_ERR_FRAM_W

Field DIS_DROP_TCPIP_ERR_FRAM writer - When this bit is set the MAC does not drop the frames which only have errors detected by the Receive Checksum engine.When this bit is reset all error frames are dropped if the Fwd_Err_Frame bit is reset.

DIS_FLUSH_RECV_FRAMES_R

Field DIS_FLUSH_RECV_FRAMES reader - When this bit is set the Rx DMA does not flush any frames because of the unavailability of receive descriptors or buffers.

DIS_FLUSH_RECV_FRAMES_W

Field DIS_FLUSH_RECV_FRAMES writer - When this bit is set the Rx DMA does not flush any frames because of the unavailability of receive descriptors or buffers.

DROP_GFRM_R

Field DROP_GFRM reader - When set the MAC drops the received giant frames in the Rx FIFO that is frames that are larger than the computed giant frame limit.

DROP_GFRM_W

Field DROP_GFRM writer - When set the MAC drops the received giant frames in the Rx FIFO that is frames that are larger than the computed giant frame limit.

FLUSH_TX_FIFO_R

Field FLUSH_TX_FIFO reader - When this bit is set the transmit FIFO controller logic is reset to its default values and thus all data in the Tx FIFO is lost or flushed. This bit is cleared internally when the flushing operation is complete.

FLUSH_TX_FIFO_W

Field FLUSH_TX_FIFO writer - When this bit is set the transmit FIFO controller logic is reset to its default values and thus all data in the Tx FIFO is lost or flushed. This bit is cleared internally when the flushing operation is complete.

FWD_ERR_FRAME_R

Field FWD_ERR_FRAME reader - When this bit is reset the Rx FIFO drops frames with error status (CRC error collision error giant frame watchdog timeout or overflow).

FWD_ERR_FRAME_W

Field FWD_ERR_FRAME writer - When this bit is reset the Rx FIFO drops frames with error status (CRC error collision error giant frame watchdog timeout or overflow).

FWD_UNDER_GF_R

Field FWD_UNDER_GF reader - When set the Rx FIFO forwards Undersized frames (that is frames with no Error and length less than 64 bytes) including pad-bytes and CRC.

FWD_UNDER_GF_W

Field FWD_UNDER_GF writer - When set the Rx FIFO forwards Undersized frames (that is frames with no Error and length less than 64 bytes) including pad-bytes and CRC.

OPT_SECOND_FRAME_R

Field OPT_SECOND_FRAME reader - When this bit is set it instructs the DMA to process the second frame of the Transmit data even before the status for the first frame is obtained.

OPT_SECOND_FRAME_W

Field OPT_SECOND_FRAME writer - When this bit is set it instructs the DMA to process the second frame of the Transmit data even before the status for the first frame is obtained.

R

Register DMAOPERATION_MODE reader

RX_STORE_FORWARD_R

Field RX_STORE_FORWARD reader - When this bit is set the MTL reads a frame from the Rx FIFO only after the complete frame has been written to it.

RX_STORE_FORWARD_W

Field RX_STORE_FORWARD writer - When this bit is set the MTL reads a frame from the Rx FIFO only after the complete frame has been written to it.

RX_THRESH_CTRL_R

Field RX_THRESH_CTRL reader - These two bits control the threshold level of the MTL Receive FIFO. Transfer (request) to DMA starts when the frame size within the MTL Receive FIFO is larger than the threshold. 2’b00: 64， 2’b01: 32， 2’b10: 96， 2’b11: 128 .

RX_THRESH_CTRL_W

Field RX_THRESH_CTRL writer - These two bits control the threshold level of the MTL Receive FIFO. Transfer (request) to DMA starts when the frame size within the MTL Receive FIFO is larger than the threshold. 2’b00: 64， 2’b01: 32， 2’b10: 96， 2’b11: 128 .

START_STOP_RX_R

Field START_STOP_RX reader - When this bit is set the Receive process is placed in the Running state. The DMA attempts to acquire the descriptor from the Receive list and processes the incoming frames.When this bit is cleared the Rx DMA operation is stopped after the transfer of the current frame.

START_STOP_RX_W

Field START_STOP_RX writer - When this bit is set the Receive process is placed in the Running state. The DMA attempts to acquire the descriptor from the Receive list and processes the incoming frames.When this bit is cleared the Rx DMA operation is stopped after the transfer of the current frame.

START_STOP_TRANSMISSION_COMMAND_R

Field START_STOP_TRANSMISSION_COMMAND reader - When this bit is set transmission is placed in the Running state and the DMA checks the Transmit List at the current position for a frame to be transmitted.When this bit is reset the transmission process is placed in the Stopped state after completing the transmission of the current frame.

START_STOP_TRANSMISSION_COMMAND_W

Field START_STOP_TRANSMISSION_COMMAND writer - When this bit is set transmission is placed in the Running state and the DMA checks the Transmit List at the current position for a frame to be transmitted.When this bit is reset the transmission process is placed in the Stopped state after completing the transmission of the current frame.

TX_STR_FWD_R

Field TX_STR_FWD reader - When this bit is set transmission starts when a full frame resides in the MTL Transmit FIFO. When this bit is set the Tx_Thresh_Ctrl values specified in Tx_Thresh_Ctrl are ignored.

TX_STR_FWD_W

Field TX_STR_FWD writer - When this bit is set transmission starts when a full frame resides in the MTL Transmit FIFO. When this bit is set the Tx_Thresh_Ctrl values specified in Tx_Thresh_Ctrl are ignored.

TX_THRESH_CTRL_R

Field TX_THRESH_CTRL reader - These bits control the threshold level of the MTL Transmit FIFO. Transmission starts when the frame size within the MTL Transmit FIFO is larger than the threshold. In addition full frames with a length less than the threshold are also transmitted. These bits are used only when Tx_Str_fwd is reset. 3’b000: 64 3’b001: 128 3’b010: 192 3’b011: 256 3’b100: 40 3’b101: 32 3’b110: 24 3’b111: 16 .

TX_THRESH_CTRL_W

Field TX_THRESH_CTRL writer - These bits control the threshold level of the MTL Transmit FIFO. Transmission starts when the frame size within the MTL Transmit FIFO is larger than the threshold. In addition full frames with a length less than the threshold are also transmitted. These bits are used only when Tx_Str_fwd is reset. 3’b000: 64 3’b001: 128 3’b010: 192 3’b011: 256 3’b100: 40 3’b101: 32 3’b110: 24 3’b111: 16 .

W

Register DMAOPERATION_MODE writer