Type Alias esp32::emac_dma::dmaoperation_mode::W

pub type W = W<DMAOPERATION_MODE_SPEC>;

Register DMAOPERATION_MODE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn start_stop_rx(&mut self) -> START_STOP_RX_W<'_, DMAOPERATION_MODE_SPEC>

Bit 1 - When this bit is set the Receive process is placed in the Running state. The DMA attempts to acquire the descriptor from the Receive list and processes the incoming frames.When this bit is cleared the Rx DMA operation is stopped after the transfer of the current frame.

pub fn opt_second_frame( &mut self ) -> OPT_SECOND_FRAME_W<'_, DMAOPERATION_MODE_SPEC>

Bit 2 - When this bit is set it instructs the DMA to process the second frame of the Transmit data even before the status for the first frame is obtained.

pub fn rx_thresh_ctrl(&mut self) -> RX_THRESH_CTRL_W<'_, DMAOPERATION_MODE_SPEC>

Bits 3:4 - These two bits control the threshold level of the MTL Receive FIFO. Transfer (request) to DMA starts when the frame size within the MTL Receive FIFO is larger than the threshold. 2’b00: 64， 2’b01: 32， 2’b10: 96， 2’b11: 128 .

pub fn drop_gfrm(&mut self) -> DROP_GFRM_W<'_, DMAOPERATION_MODE_SPEC>

Bit 5 - When set the MAC drops the received giant frames in the Rx FIFO that is frames that are larger than the computed giant frame limit.

pub fn fwd_under_gf(&mut self) -> FWD_UNDER_GF_W<'_, DMAOPERATION_MODE_SPEC>

Bit 6 - When set the Rx FIFO forwards Undersized frames (that is frames with no Error and length less than 64 bytes) including pad-bytes and CRC.

pub fn fwd_err_frame(&mut self) -> FWD_ERR_FRAME_W<'_, DMAOPERATION_MODE_SPEC>

Bit 7 - When this bit is reset the Rx FIFO drops frames with error status (CRC error collision error giant frame watchdog timeout or overflow).

pub fn start_stop_transmission_command( &mut self ) -> START_STOP_TRANSMISSION_COMMAND_W<'_, DMAOPERATION_MODE_SPEC>

Bit 13 - When this bit is set transmission is placed in the Running state and the DMA checks the Transmit List at the current position for a frame to be transmitted.When this bit is reset the transmission process is placed in the Stopped state after completing the transmission of the current frame.

pub fn tx_thresh_ctrl(&mut self) -> TX_THRESH_CTRL_W<'_, DMAOPERATION_MODE_SPEC>

Bits 14:16 - These bits control the threshold level of the MTL Transmit FIFO. Transmission starts when the frame size within the MTL Transmit FIFO is larger than the threshold. In addition full frames with a length less than the threshold are also transmitted. These bits are used only when Tx_Str_fwd is reset. 3’b000: 64 3’b001: 128 3’b010: 192 3’b011: 256 3’b100: 40 3’b101: 32 3’b110: 24 3’b111: 16 .

pub fn flush_tx_fifo(&mut self) -> FLUSH_TX_FIFO_W<'_, DMAOPERATION_MODE_SPEC>

Bit 20 - When this bit is set the transmit FIFO controller logic is reset to its default values and thus all data in the Tx FIFO is lost or flushed. This bit is cleared internally when the flushing operation is complete.

pub fn tx_str_fwd(&mut self) -> TX_STR_FWD_W<'_, DMAOPERATION_MODE_SPEC>

Bit 21 - When this bit is set transmission starts when a full frame resides in the MTL Transmit FIFO. When this bit is set the Tx_Thresh_Ctrl values specified in Tx_Thresh_Ctrl are ignored.

pub fn dis_flush_recv_frames( &mut self ) -> DIS_FLUSH_RECV_FRAMES_W<'_, DMAOPERATION_MODE_SPEC>

Bit 24 - When this bit is set the Rx DMA does not flush any frames because of the unavailability of receive descriptors or buffers.

pub fn rx_store_forward( &mut self ) -> RX_STORE_FORWARD_W<'_, DMAOPERATION_MODE_SPEC>

Bit 25 - When this bit is set the MTL reads a frame from the Rx FIFO only after the complete frame has been written to it.

pub fn dis_drop_tcpip_err_fram( &mut self ) -> DIS_DROP_TCPIP_ERR_FRAM_W<'_, DMAOPERATION_MODE_SPEC>

Bit 26 - When this bit is set the MAC does not drop the frames which only have errors detected by the Receive Checksum engine.When this bit is reset all error frames are dropped if the Fwd_Err_Frame bit is reset.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual