Module esp32::emac_mac::emacaddr2high

Upper 16 bits of the third 6-byte MAC address

Structs

EMACADDR2HIGH_SPEC

Upper 16 bits of the third 6-byte MAC address

Type Aliases

ADDRESS_ENABLE2_R

Field ADDRESS_ENABLE2 reader - When this bit is set the address filter module uses the third MAC address for perfect filtering. When this bit is reset the address filter module ignores the address for filtering.

ADDRESS_ENABLE2_W

Field ADDRESS_ENABLE2 writer - When this bit is set the address filter module uses the third MAC address for perfect filtering. When this bit is reset the address filter module ignores the address for filtering.

MAC_ADDRESS2_HI_R

Field MAC_ADDRESS2_HI reader - This field contains the upper 16 bits Bits[47:32] of the third 6-byte MAC address.

MAC_ADDRESS2_HI_W

Field MAC_ADDRESS2_HI writer - This field contains the upper 16 bits Bits[47:32] of the third 6-byte MAC address.

MASK_BYTE_CONTROL2_R

Field MASK_BYTE_CONTROL2 reader - These bits are mask control bits for comparison of each of the EMACADDR2 bytes. When set high the MAC does not compare the corresponding byte of received DA or SA with the contents of EMACADDR2 registers. Each bit controls the masking of the bytes as follows: Bit[29]: EMACADDR2 High [15:8]. Bit[28]: EMACADDR2 High [7:0]. Bit[27]: EMACADDR2 Low [31:24]. Bit[24]: EMACADDR2 Low [7:0].You can filter a group of addresses (known as group address filtering) by masking one or more bytes of the address.

MASK_BYTE_CONTROL2_W

Field MASK_BYTE_CONTROL2 writer - These bits are mask control bits for comparison of each of the EMACADDR2 bytes. When set high the MAC does not compare the corresponding byte of received DA or SA with the contents of EMACADDR2 registers. Each bit controls the masking of the bytes as follows: Bit[29]: EMACADDR2 High [15:8]. Bit[28]: EMACADDR2 High [7:0]. Bit[27]: EMACADDR2 Low [31:24]. Bit[24]: EMACADDR2 Low [7:0].You can filter a group of addresses (known as group address filtering) by masking one or more bytes of the address.

R

Register EMACADDR2HIGH reader

SOURCE_ADDRESS2_R

Field SOURCE_ADDRESS2 reader - When this bit is set the EMACADDR2[47:0] is used to compare with the SA fields of the received frame. When this bit is reset the EMACADDR2[47:0] is used to compare with the DA fields of the received frame.

SOURCE_ADDRESS2_W

Field SOURCE_ADDRESS2 writer - When this bit is set the EMACADDR2[47:0] is used to compare with the SA fields of the received frame. When this bit is reset the EMACADDR2[47:0] is used to compare with the DA fields of the received frame.

W

Register EMACADDR2HIGH writer