Type Alias esp32::emac_mac::emacaddr3high::MASK_BYTE_CONTROL3_R

pub type MASK_BYTE_CONTROL3_R = FieldReader;

Field MASK_BYTE_CONTROL3 reader - These bits are mask control bits for comparison of each of the EMACADDR3 bytes. When set high the MAC does not compare the corresponding byte of received DA or SA with the contents of EMACADDR3 registers. Each bit controls the masking of the bytes as follows: Bit[29]: EMACADDR3 High [15:8]. Bit[28]: EMACADDR3 High [7:0]. Bit[27]: EMACADDR3 Low [31:24]. Bit[24]: EMACADDR3 Low [7:0].You can filter a group of addresses (known as group address filtering) by masking one or more bytes of the address.

Aliased Type

struct MASK_BYTE_CONTROL3_R { /* private fields */ }