Type Alias esp32::emac_mac::emacaddr7high::R

pub type R = R<EMACADDR7HIGH_SPEC>;

Register EMACADDR7HIGH reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn mac_address7_hi(&self) -> MAC_ADDRESS7_HI_R

Bits 0:15 - This field contains the upper 16 bits Bits[47:32] of the eighth 6-byte MAC Address.

pub fn mask_byte_control7(&self) -> MASK_BYTE_CONTROL7_R

Bits 24:29 - These bits are mask control bits for comparison of each of the EMACADDR7 bytes. When set high the MAC does not compare the corresponding byte of received DA or SA with the contents of EMACADDR7 registers. Each bit controls the masking of the bytes as follows: Bit[29]: EMACADDR7 High [15:8]. Bit[28]: EMACADDR7 High [7:0]. Bit[27]: EMACADDR7 Low [31:24]. Bit[24]: EMACADDR7 Low [7:0].You can filter a group of addresses (known as group address filtering) by masking one or more bytes of the address.

pub fn source_address7(&self) -> SOURCE_ADDRESS7_R

Bit 30 - When this bit is set the EMACADDR7[47:0] is used to compare with the SA fields of the received frame. When this bit is reset the EMACADDR7[47:0] is used to compare with the DA fields of the received frame.

pub fn address_enable7(&self) -> ADDRESS_ENABLE7_R

Bit 31 - When this bit is set the address filter module uses the eighth MAC address for perfect filtering. When this bit is reset the address filter module ignores the address for filtering.