Type Alias esp32::emac_mac::emacfc::PLT_W

pub type PLT_W<'a, REG> = FieldWriter<'a, REG, 2>;

Field PLT writer - This field configures the threshold of the Pause timer automatic retransmission of the Pause frame.The threshold values should be always less than the Pause Time configured in Bits[31:16]. For example if PT = 100H (256 slot-times) and PLT = 01 then a second Pause frame is automatically transmitted at 228 (256-28) slot times after the first Pause frame is transmitted. The following list provides the threshold values for different values: 2’b00: The threshold is Pause time minus 4 slot times (PT-4 slot times). 2’b01: The threshold is Pause time minus 28 slot times (PT-28 slot times). 2’b10: The threshold is Pause time minus 144 slot times (PT-144 slot times). 2’b11: The threshold is Pause time minus 256 slot times (PT-256 slot times). The slot time is defined as the time taken to transmit 512 bits (64 bytes) on the MII interface.

Aliased Type

struct PLT_W<'a, REG> { /* private fields */ }