Type Alias esp32::emac_mac::emacff::R

pub type R = R<EMACFF_SPEC>;

Register EMACFF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn pmode(&self) -> PMODE_R

Bit 0 - When this bit is set the Address Filter module passes all incoming frames irrespective of the destination or source address. The SA or DA Filter Fails status bits of the Receive Status Word are always cleared when PR(PRI_RATIO) is set.

pub fn daif(&self) -> DAIF_R

Bit 3 - When this bit is set the Address Check block operates in inverse filtering mode for the DA address comparison for both unicast and multicast frames. When reset normal filtering of frames is performed.

pub fn pam(&self) -> PAM_R

Bit 4 - When set this bit indicates that all received frames with a multicast destination address (first bit in the destination address field is ‘1’) are passed.

pub fn dbf(&self) -> DBF_R

Bit 5 - When this bit is set the AFM(Address Filtering Module) module blocks all incoming broadcast frames. In addition it overrides all other filter settings. When this bit is reset the AFM module passes all received broadcast Frames.

pub fn pcf(&self) -> PCF_R

Bits 6:7 - These bits control the forwarding of all control frames (including unicast and multicast Pause frames). 2’b00: MAC filters all control frames from reaching the application. 2’b01: MAC forwards all control frames except Pause frames to application even if they fail the Address filter. 2’b10: MAC forwards all control frames to application even if they fail the Address Filter. 2’b11: MAC forwards control frames that pass the Address Filter.The following conditions should be true for the Pause frames processing: Condition 1: The MAC is in the full-duplex mode and flow control is enabled by setting Bit 2 (RFE) of Register (Flow Control Register) to 1. Condition 2: The destination address (DA) of the received frame matches the special multicast address or the MAC Address 0 when Bit 3 (UP) of the Register(Flow Control Register) is set. Condition 3: The Type field of the received frame is 0x8808 and the OPCODE field is 0x0001.

pub fn saif(&self) -> SAIF_R

Bit 8 - When this bit is set the Address Check block operates in inverse filtering mode for the SA address comparison. The frames whose SA matches the SA registers are marked as failing the SA Address filter. When this bit is reset frames whose SA does not match the SA registers are marked as failing the SA Address filter.

pub fn safe(&self) -> SAFE_R

Bit 9 - When this bit is set the MAC compares the SA field of the received frames with the values programmed in the enabled SA registers. If the comparison fails the MAC drops the frame. When this bit is reset the MAC forwards the received frame to the application with updated SAF bit of the Rx Status depending on the SA address comparison.

pub fn receive_all(&self) -> RECEIVE_ALL_R

Bit 31 - When this bit is set the MAC Receiver module passes all received frames irrespective of whether they pass the address filter or not to the Application. The result of the SA or DA filtering is updated (pass or fail) in the corresponding bits in the Receive Status Word. When this bit is reset the Receiver module passes only those frames to the Application that pass the SA or DA address Filter.