Type Alias esp32::emac_mac::emaclpi_crs::R

pub type R = R<EMACLPI_CRS_SPEC>;

Register EMACLPI_CRS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn tlpien(&self) -> TLPIEN_R

Bit 0 - When set this bit indicates that the MAC Transmitter has entered the LPI state because of the setting of the LPIEN bit. This bit is cleared by a read into this register.

pub fn tlpiex(&self) -> TLPIEX_R

Bit 1 - When set this bit indicates that the MAC transmitter has exited the LPI state after the user has cleared the LPIEN bit and the LPI_TW_Timer has expired.This bit is cleared by a read into this register.

pub fn rlpien(&self) -> RLPIEN_R

Bit 2 - When set this bit indicates that the MAC Receiver has received an LPI pattern and entered the LPI state. This bit is cleared by a read into this register.

pub fn rlpiex(&self) -> RLPIEX_R

Bit 3 - When set this bit indicates that the MAC Receiver has stopped receiving the LPI pattern on the MII interface exited the LPI state and resumed the normal reception. This bit is cleared by a read into this register.

pub fn tlpist(&self) -> TLPIST_R

Bit 8 - When set this bit indicates that the MAC is transmitting the LPI pattern on the MII interface.

pub fn rlpist(&self) -> RLPIST_R

Bit 9 - When set this bit indicates that the MAC is receiving the LPI pattern on the MII interface.

pub fn lpien(&self) -> LPIEN_R

Bit 16 - When set this bit instructs the MAC Transmitter to enter the LPI state. When reset this bit instructs the MAC to exit the LPI state and resume normal transmission.This bit is cleared when the LPITXA bit is set and the MAC exits the LPI state because of the arrival of a new packet for transmission.

pub fn pls(&self) -> PLS_R

Bit 17 - This bit indicates the link status of the PHY.When set the link is considered to be okay (up) and when reset the link is considered to be down.

pub fn lpitxa(&self) -> LPITXA_R

Bit 19 - This bit controls the behavior of the MAC when it is entering or coming out of the LPI mode on the transmit side.If the LPITXA and LPIEN bits are set to 1 the MAC enters the LPI mode only after all outstanding frames and pending frames have been transmitted. The MAC comes out of the LPI mode when the application sends any frame.When this bit is 0 the LPIEN bit directly controls behavior of the MAC when it is entering or coming out of the LPI mode.