Module esp32::emac_mac::pmt_csr

PMT Control and Status

Structs

PMT_CSR_SPEC

PMT Control and Status

Type Aliases

GLBLUCAST_R

Field GLBLUCAST reader - When set enables any unicast packet filtered by the MAC (DAFilter) address recognition to be a remote wake-up frame.

MGKPKTEN_R

Field MGKPKTEN reader - When set enables generation of a power management event because of magic packet reception.

MGKPRCVD_R

Field MGKPRCVD reader - When set this bit indicates that the power management event is generated because of the reception of a magic packet. This bit is cleared by a Read into this register.

PWRDWN_R

Field PWRDWN reader - When set the MAC receiver drops all received frames until it receives the expected magic packet or remote wake-up frame.This bit must only be set when MGKPKTEN GLBLUCAST or RWKPKTEN bit is set high.

R

Register PMT_CSR reader

RWKFILTRST_R

Field RWKFILTRST reader - When this bit is set it resets the RWKPTR register to 3’b000.

RWKPKTEN_R

Field RWKPKTEN reader - When set enables generation of a power management event because of remote wake-up frame reception

RWKPRCVD_R

Field RWKPRCVD reader - When set this bit indicates the power management event is generated because of the reception of a remote wake-up frame. This bit is cleared by a Read into this register.

RWKPTR_R

Field RWKPTR reader - The maximum value of the pointer is 7 the detail information please refer to PMT_RWUFFR.