Type Alias esp32::emac_mac::pmt_csr::R

pub type R = R<PMT_CSR_SPEC>;

Register PMT_CSR reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn pwrdwn(&self) -> PWRDWN_R

Bit 0 - When set the MAC receiver drops all received frames until it receives the expected magic packet or remote wake-up frame.This bit must only be set when MGKPKTEN GLBLUCAST or RWKPKTEN bit is set high.

pub fn mgkpkten(&self) -> MGKPKTEN_R

Bit 1 - When set enables generation of a power management event because of magic packet reception.

pub fn rwkpkten(&self) -> RWKPKTEN_R

Bit 2 - When set enables generation of a power management event because of remote wake-up frame reception

pub fn mgkprcvd(&self) -> MGKPRCVD_R

Bit 5 - When set this bit indicates that the power management event is generated because of the reception of a magic packet. This bit is cleared by a Read into this register.

pub fn rwkprcvd(&self) -> RWKPRCVD_R

Bit 6 - When set this bit indicates the power management event is generated because of the reception of a remote wake-up frame. This bit is cleared by a Read into this register.

pub fn glblucast(&self) -> GLBLUCAST_R

Bit 9 - When set enables any unicast packet filtered by the MAC (DAFilter) address recognition to be a remote wake-up frame.

pub fn rwkptr(&self) -> RWKPTR_R

Bits 24:28 - The maximum value of the pointer is 7 the detail information please refer to PMT_RWUFFR.

pub fn rwkfiltrst(&self) -> RWKFILTRST_R

Bit 31 - When this bit is set it resets the RWKPTR register to 3’b000.