Module esp32::i2c0::sr

Structs

SR_SPEC

You can read this register and get sr::R. See API.

Type Aliases

ACK_REC_R

Field ACK_REC reader - This register stores the value of ACK bit.

ARB_LOST_R

Field ARB_LOST reader - when I2C lost control of SDA line this register changes to high level.

BUS_BUSY_R

Field BUS_BUSY reader - 1:I2C bus is busy transferring data. 0:I2C bus is in idle state.

BYTE_TRANS_R

Field BYTE_TRANS reader - This register changes to high level when one byte is transferred.

R

Register SR reader

RXFIFO_CNT_R

Field RXFIFO_CNT reader - This register represent the amount of data need to send.

SCL_MAIN_STATE_LAST_R

Field SCL_MAIN_STATE_LAST reader - This register stores the value of state machine for i2c module. 3’h0: SCL_MAIN_IDLE 3’h1: SCL_ADDRESS_SHIFT 3’h2: SCL_ACK_ADDRESS 3’h3: SCL_RX_DATA 3’h4 SCL_TX_DATA 3’h5:SCL_SEND_ACK 3’h6:SCL_WAIT_ACK

SCL_STATE_LAST_R

Field SCL_STATE_LAST reader - This register stores the value of state machine to produce SCL. 3’h0: SCL_IDLE 3’h1:SCL_START 3’h2:SCL_LOW_EDGE 3’h3: SCL_LOW 3’h4:SCL_HIGH_EDGE 3’h5:SCL_HIGH 3’h6:SCL_STOP

SLAVE_ADDRESSED_R

Field SLAVE_ADDRESSED reader - when configured as i2c slave and the address send by master is equal to slave’s address then this bit will be high level.

SLAVE_RW_R

Field SLAVE_RW reader - when in slave mode 1: master read slave 0: master write slave.

TIME_OUT_R

Field TIME_OUT reader - when I2C takes more than time_out_reg clocks to receive a data then this register changes to high level.

TXFIFO_CNT_R

Field TXFIFO_CNT reader - This register stores the amount of received data in ram.