Module esp32::io_mux::gpio16

Structs

GPIO16_SPEC

You can read this register and get gpio16::R. You can reset, write, write_with_zero this register using gpio16::W. You can also modify this register. See API.

Type Aliases

FUN_DRV_R

Field FUN_DRV reader - Select the drive strength of the pad. A higher value corresponds with a higher strength. For GPIO34-39, FUN_DRV is always 0. For detailed drive strength, please see note 8 in Table ”Notes on ESP32 Pin Lists”, in ESP32 Datasheet.

FUN_DRV_W

Field FUN_DRV writer - Select the drive strength of the pad. A higher value corresponds with a higher strength. For GPIO34-39, FUN_DRV is always 0. For detailed drive strength, please see note 8 in Table ”Notes on ESP32 Pin Lists”, in ESP32 Datasheet.

FUN_IE_R

Field FUN_IE reader - Input enable of the pad. 1: input enabled; 0: input disabled.

FUN_IE_W

Field FUN_IE writer - Input enable of the pad. 1: input enabled; 0: input disabled.

FUN_WPD_R

Field FUN_WPD reader - Pull-down enable of the pad. 1: internal pull-down enabled, 0: internal pull-down disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull-down circuitry, therefore, their FUN_WPD is always 0.

FUN_WPD_W

Field FUN_WPD writer - Pull-down enable of the pad. 1: internal pull-down enabled, 0: internal pull-down disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull-down circuitry, therefore, their FUN_WPD is always 0.

FUN_WPU_R

Field FUN_WPU reader - Pull-up enable of the pad. 1: internal pull-up enabled; 0: internal pull-up disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull- down circuitry, therefore, their FUN_WPU is always 0.

FUN_WPU_W

Field FUN_WPU writer - Pull-up enable of the pad. 1: internal pull-up enabled; 0: internal pull-up disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull- down circuitry, therefore, their FUN_WPU is always 0.

MCU_DRV_R

Field MCU_DRV reader - Select the drive strength of the pad during sleep mode. A higher value corresponds with a higher strength.

MCU_DRV_W

Field MCU_DRV writer - Select the drive strength of the pad during sleep mode. A higher value corresponds with a higher strength.

MCU_IE_R

Field MCU_IE reader - Input enable of the pad during sleep mode. 1: input enabled; 0: input disabled.

MCU_IE_W

Field MCU_IE writer - Input enable of the pad during sleep mode. 1: input enabled; 0: input disabled.

MCU_OE_R

Field MCU_OE reader - Output enable of the pad in sleep mode. 1: enable output; 0: disable output.

MCU_OE_W

Field MCU_OE writer - Output enable of the pad in sleep mode. 1: enable output; 0: disable output.

MCU_SEL_R

Field MCU_SEL reader - Select the IO_MUX function for this signal. 0 selects Function 0, 1 selects Function 1, etc.

MCU_SEL_W

Field MCU_SEL writer - Select the IO_MUX function for this signal. 0 selects Function 0, 1 selects Function 1, etc.

MCU_WPD_R

Field MCU_WPD reader - Pull-down enable of the pad during sleep mode. 1: internal pull-down enabled; 0: internal pull-down disabled.

MCU_WPD_W

Field MCU_WPD writer - Pull-down enable of the pad during sleep mode. 1: internal pull-down enabled; 0: internal pull-down disabled.

MCU_WPU_R

Field MCU_WPU reader - Pull-up enable of the pad during sleep mode. 1: internal pull-up enabled; 0: internal pull-up disabled.

MCU_WPU_W

Field MCU_WPU writer - Pull-up enable of the pad during sleep mode. 1: internal pull-up enabled; 0: internal pull-up disabled.

R

Register GPIO16 reader

SLP_SEL_R

Field SLP_SEL reader - Sleep mode selection of this pad. Set to 1 to put the pad in sleep mode.

SLP_SEL_W

Field SLP_SEL writer - Sleep mode selection of this pad. Set to 1 to put the pad in sleep mode.

W

Register GPIO16 writer