Type Alias esp32::io_mux::gpio19::R

pub type R = R<GPIO19_SPEC>;

Register GPIO19 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn mcu_oe(&self) -> MCU_OE_R

Bit 0 - Output enable of the pad in sleep mode. 1: enable output; 0: disable output.

pub fn slp_sel(&self) -> SLP_SEL_R

Bit 1 - Sleep mode selection of this pad. Set to 1 to put the pad in sleep mode.

pub fn mcu_wpd(&self) -> MCU_WPD_R

Bit 2 - Pull-down enable of the pad during sleep mode. 1: internal pull-down enabled; 0: internal pull-down disabled.

pub fn mcu_wpu(&self) -> MCU_WPU_R

Bit 3 - Pull-up enable of the pad during sleep mode. 1: internal pull-up enabled; 0: internal pull-up disabled.

pub fn mcu_ie(&self) -> MCU_IE_R

Bit 4 - Input enable of the pad during sleep mode. 1: input enabled; 0: input disabled.

pub fn mcu_drv(&self) -> MCU_DRV_R

Bits 5:6 - Select the drive strength of the pad during sleep mode. A higher value corresponds with a higher strength.

pub fn fun_wpd(&self) -> FUN_WPD_R

Bit 7 - Pull-down enable of the pad. 1: internal pull-down enabled, 0: internal pull-down disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull-down circuitry, therefore, their FUN_WPD is always 0.

pub fn fun_wpu(&self) -> FUN_WPU_R

Bit 8 - Pull-up enable of the pad. 1: internal pull-up enabled; 0: internal pull-up disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull- down circuitry, therefore, their FUN_WPU is always 0.

pub fn fun_ie(&self) -> FUN_IE_R

Bit 9 - Input enable of the pad. 1: input enabled; 0: input disabled.

pub fn fun_drv(&self) -> FUN_DRV_R

Bits 10:11 - Select the drive strength of the pad. A higher value corresponds with a higher strength. For GPIO34-39, FUN_DRV is always 0. For detailed drive strength, please see note 8 in Table ”Notes on ESP32 Pin Lists”, in ESP32 Datasheet.

pub fn mcu_sel(&self) -> MCU_SEL_R

Bits 12:14 - Select the IO_MUX function for this signal. 0 selects Function 0, 1 selects Function 1, etc.