Type Alias esp32::io_mux::gpio8::W

pub type W = W<GPIO8_SPEC>;

Register GPIO8 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn mcu_oe(&mut self) -> MCU_OE_W<'_, GPIO8_SPEC>

Bit 0 - Output enable of the pad in sleep mode. 1: enable output; 0: disable output.

pub fn slp_sel(&mut self) -> SLP_SEL_W<'_, GPIO8_SPEC>

Bit 1 - Sleep mode selection of this pad. Set to 1 to put the pad in sleep mode.

pub fn mcu_wpd(&mut self) -> MCU_WPD_W<'_, GPIO8_SPEC>

Bit 2 - Pull-down enable of the pad during sleep mode. 1: internal pull-down enabled; 0: internal pull-down disabled.

pub fn mcu_wpu(&mut self) -> MCU_WPU_W<'_, GPIO8_SPEC>

Bit 3 - Pull-up enable of the pad during sleep mode. 1: internal pull-up enabled; 0: internal pull-up disabled.

pub fn mcu_ie(&mut self) -> MCU_IE_W<'_, GPIO8_SPEC>

Bit 4 - Input enable of the pad during sleep mode. 1: input enabled; 0: input disabled.

pub fn mcu_drv(&mut self) -> MCU_DRV_W<'_, GPIO8_SPEC>

Bits 5:6 - Select the drive strength of the pad during sleep mode. A higher value corresponds with a higher strength.

pub fn fun_wpd(&mut self) -> FUN_WPD_W<'_, GPIO8_SPEC>

Bit 7 - Pull-down enable of the pad. 1: internal pull-down enabled, 0: internal pull-down disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull-down circuitry, therefore, their FUN_WPD is always 0.

pub fn fun_wpu(&mut self) -> FUN_WPU_W<'_, GPIO8_SPEC>

Bit 8 - Pull-up enable of the pad. 1: internal pull-up enabled; 0: internal pull-up disabled. GPIO pins 34-39 are input-only. These pins do not feature an output driver or internal pull- up/pull- down circuitry, therefore, their FUN_WPU is always 0.

pub fn fun_ie(&mut self) -> FUN_IE_W<'_, GPIO8_SPEC>

Bit 9 - Input enable of the pad. 1: input enabled; 0: input disabled.

pub fn fun_drv(&mut self) -> FUN_DRV_W<'_, GPIO8_SPEC>

Bits 10:11 - Select the drive strength of the pad. A higher value corresponds with a higher strength. For GPIO34-39, FUN_DRV is always 0. For detailed drive strength, please see note 8 in Table ”Notes on ESP32 Pin Lists”, in ESP32 Datasheet.

pub fn mcu_sel(&mut self) -> MCU_SEL_W<'_, GPIO8_SPEC>

Bits 12:14 - Select the IO_MUX function for this signal. 0 selects Function 0, 1 selects Function 1, etc.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual