Type Alias esp32::ledc::hstimer_conf::R

pub type R = R<HSTIMER_CONF_SPEC>;

Register HSTIMER%s_CONF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn duty_res(&self) -> DUTY_RES_R

Bits 0:4 - This register controls the range of the counter in high speed timer0. the counter range is [0 2**reg_hstimer0_lim] the max bit width for counter is 20.

pub fn div_num(&self) -> DIV_NUM_R

Bits 5:22 - This register is used to configure parameter for divider in high speed timer0 the least significant eight bits represent the decimal part.

pub fn pause(&self) -> PAUSE_R

Bit 23 - This bit is used to pause the counter in high speed timer0

pub fn rst(&self) -> RST_R

Bit 24 - This bit is used to reset high speed timer0 the counter will be 0 after reset.

pub fn tick_sel(&self) -> TICK_SEL_R

Bit 25 - This bit is used to choose apb_clk or ref_tick for high speed timer0. 1’b1:apb_clk 0:ref_tick