Type Alias esp32::ledc::lsch_conf0::R

pub type R = R<LSCH_CONF0_SPEC>;

Register LSCH%s_CONF0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn timer_sel(&self) -> TIMER_SEL_R

Bits 0:1 - There are four low speed timers the two bits are used to select one of them for low speed channel0. 2’b00: seletc lstimer0. 2’b01: select lstimer1. 2’b10: select lstimer2. 2’b11: select lstimer3.

pub fn sig_out_en(&self) -> SIG_OUT_EN_R

Bit 2 - This is the output enable control bit for low speed channel0.

pub fn idle_lv(&self) -> IDLE_LV_R

Bit 3 - This bit is used to control the output value when low speed channel0 is off.

pub fn para_up(&self) -> PARA_UP_R

Bit 4 - This bit is used to update register LEDC_LSCH0_HPOINT and LEDC_LSCH0_DUTY for low speed channel0.