Module esp32::pcnt::u_conf0

Structs

U_CONF0_SPEC

You can read this register and get u_conf0::R. You can reset, write, write_with_zero this register using u_conf0::W. You can also modify this register. See API.

Type Aliases

CH0_HCTRL_MODE_R

Field CH0_HCTRL_MODE reader - This register is used to control the mode of channel0’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH0_HCTRL_MODE_W

Field CH0_HCTRL_MODE writer - This register is used to control the mode of channel0’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH0_LCTRL_MODE_R

Field CH0_LCTRL_MODE reader - This register is used to control the mode of channel0’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH0_LCTRL_MODE_W

Field CH0_LCTRL_MODE writer - This register is used to control the mode of channel0’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH0_NEG_MODE_R

Field CH0_NEG_MODE reader - This register is used to control the mode of channel0’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

CH0_NEG_MODE_W

Field CH0_NEG_MODE writer - This register is used to control the mode of channel0’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

CH0_POS_MODE_R

Field CH0_POS_MODE reader - This register is used to control the mode of channel0’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

CH0_POS_MODE_W

Field CH0_POS_MODE writer - This register is used to control the mode of channel0’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

CH1_HCTRL_MODE_R

Field CH1_HCTRL_MODE reader - This register is used to control the mode of channel1’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH1_HCTRL_MODE_W

Field CH1_HCTRL_MODE writer - This register is used to control the mode of channel1’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH1_LCTRL_MODE_R

Field CH1_LCTRL_MODE reader - This register is used to control the mode of channel1’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH1_LCTRL_MODE_W

Field CH1_LCTRL_MODE writer - This register is used to control the mode of channel1’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

CH1_NEG_MODE_R

Field CH1_NEG_MODE reader - This register is used to control the mode of channel1’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

CH1_NEG_MODE_W

Field CH1_NEG_MODE writer - This register is used to control the mode of channel1’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

CH1_POS_MODE_R

Field CH1_POS_MODE reader - This register is used to control the mode of channel1’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

CH1_POS_MODE_W

Field CH1_POS_MODE writer - This register is used to control the mode of channel1’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

FILTER_EN_R

Field FILTER_EN reader - This is the enable bit for filtering input signals for unit0.

FILTER_EN_W

Field FILTER_EN writer - This is the enable bit for filtering input signals for unit0.

FILTER_THRES_R

Field FILTER_THRES reader - This register is used to filter pluse whose width is smaller than this value for unit0.

FILTER_THRES_W

Field FILTER_THRES writer - This register is used to filter pluse whose width is smaller than this value for unit0.

R

Register U%s_CONF0 reader

THR_H_LIM_EN_R

Field THR_H_LIM_EN reader - This is the enable bit for comparing unit0’s count with thr_h_lim value.

THR_H_LIM_EN_W

Field THR_H_LIM_EN writer - This is the enable bit for comparing unit0’s count with thr_h_lim value.

THR_L_LIM_EN_R

Field THR_L_LIM_EN reader - This is the enable bit for comparing unit0’s count with thr_l_lim value.

THR_L_LIM_EN_W

Field THR_L_LIM_EN writer - This is the enable bit for comparing unit0’s count with thr_l_lim value.

THR_THRES0_EN_R

Field THR_THRES0_EN reader - This is the enable bit for comparing unit0’s count with thres0 value.

THR_THRES0_EN_W

Field THR_THRES0_EN writer - This is the enable bit for comparing unit0’s count with thres0 value.

THR_THRES1_EN_R

Field THR_THRES1_EN reader - This is the enable bit for comparing unit0’s count with thres1 value .

THR_THRES1_EN_W

Field THR_THRES1_EN writer - This is the enable bit for comparing unit0’s count with thres1 value .

THR_ZERO_EN_R

Field THR_ZERO_EN reader - This is the enable bit for comparing unit0’s count with 0 value.

THR_ZERO_EN_W

Field THR_ZERO_EN writer - This is the enable bit for comparing unit0’s count with 0 value.

W

Register U%s_CONF0 writer