Type Alias esp32::pcnt::u_conf0::R

pub type R = R<U_CONF0_SPEC>;

Register U%s_CONF0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn filter_thres(&self) -> FILTER_THRES_R

Bits 0:9 - This register is used to filter pluse whose width is smaller than this value for unit0.

pub fn filter_en(&self) -> FILTER_EN_R

Bit 10 - This is the enable bit for filtering input signals for unit0.

pub fn thr_zero_en(&self) -> THR_ZERO_EN_R

Bit 11 - This is the enable bit for comparing unit0’s count with 0 value.

pub fn thr_h_lim_en(&self) -> THR_H_LIM_EN_R

Bit 12 - This is the enable bit for comparing unit0’s count with thr_h_lim value.

pub fn thr_l_lim_en(&self) -> THR_L_LIM_EN_R

Bit 13 - This is the enable bit for comparing unit0’s count with thr_l_lim value.

pub fn thr_thres0_en(&self) -> THR_THRES0_EN_R

Bit 14 - This is the enable bit for comparing unit0’s count with thres0 value.

pub fn thr_thres1_en(&self) -> THR_THRES1_EN_R

Bit 15 - This is the enable bit for comparing unit0’s count with thres1 value .

pub fn ch0_neg_mode(&self) -> CH0_NEG_MODE_R

Bits 16:17 - This register is used to control the mode of channel0’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

pub fn ch0_pos_mode(&self) -> CH0_POS_MODE_R

Bits 18:19 - This register is used to control the mode of channel0’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

pub fn ch0_hctrl_mode(&self) -> CH0_HCTRL_MODE_R

Bits 20:21 - This register is used to control the mode of channel0’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

pub fn ch0_lctrl_mode(&self) -> CH0_LCTRL_MODE_R

Bits 22:23 - This register is used to control the mode of channel0’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

pub fn ch1_neg_mode(&self) -> CH1_NEG_MODE_R

Bits 24:25 - This register is used to control the mode of channel1’s input negedge signal for unit0. 2’d1: increase at the negedge of input signal 2’d2:decrease at the negedge of input signal others:forbidden

pub fn ch1_pos_mode(&self) -> CH1_POS_MODE_R

Bits 26:27 - This register is used to control the mode of channel1’s input posedge signal for unit0. 2’d1: increase at the posedge of input signal 2’d2:decrease at the posedge of input signal others:forbidden

pub fn ch1_hctrl_mode(&self) -> CH1_HCTRL_MODE_R

Bits 28:29 - This register is used to control the mode of channel1’s high control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden

pub fn ch1_lctrl_mode(&self) -> CH1_LCTRL_MODE_R

Bits 30:31 - This register is used to control the mode of channel1’s low control signal for unit0. 2’d0:increase when control signal is low 2’d1: decrease when control signal is high others:forbidden