Module esp32::rmt::chconf0

Structs

CHCONF0_SPEC

You can read this register and get chconf0::R. You can reset, write, write_with_zero this register using chconf0::W. You can also modify this register. See API.

Type Aliases

CARRIER_EN_R

Field CARRIER_EN reader - This is the carrier modulation enable control bit for channel0.

CARRIER_EN_W

Field CARRIER_EN writer - This is the carrier modulation enable control bit for channel0.

CARRIER_OUT_LV_R

Field CARRIER_OUT_LV reader - This bit is used to configure the way carrier wave is modulated for channel0.1’b1:transmit on low output level 1’b0:transmit on high output level.

CARRIER_OUT_LV_W

Field CARRIER_OUT_LV writer - This bit is used to configure the way carrier wave is modulated for channel0.1’b1:transmit on low output level 1’b0:transmit on high output level.

CLK_EN_R

Field CLK_EN reader - This bit is used to control clock.when software config RMT internal registers it controls the register clock.

CLK_EN_W

Field CLK_EN writer - This bit is used to control clock.when software config RMT internal registers it controls the register clock.

DIV_CNT_R

Field DIV_CNT reader - This register is used to configure the frequency divider’s factor in channel0.

DIV_CNT_W

Field DIV_CNT writer - This register is used to configure the frequency divider’s factor in channel0.

IDLE_THRES_R

Field IDLE_THRES reader - In receive mode when no edge is detected on the input signal for longer than reg_idle_thres_ch0 then the receive process is done.

IDLE_THRES_W

Field IDLE_THRES writer - In receive mode when no edge is detected on the input signal for longer than reg_idle_thres_ch0 then the receive process is done.

MEM_PD_R

Field MEM_PD reader - This bit is used to reduce power consumed by mem. 1:mem is in low power state.

MEM_PD_W

Field MEM_PD writer - This bit is used to reduce power consumed by mem. 1:mem is in low power state.

MEM_SIZE_R

Field MEM_SIZE reader - This register is used to configure the the amount of memory blocks allocated to channel0.

MEM_SIZE_W

Field MEM_SIZE writer - This register is used to configure the the amount of memory blocks allocated to channel0.

R

Register CH%sCONF0 reader

W

Register CH%sCONF0 writer