Type Alias esp32::rmt::chconf0::R

pub type R = R<CHCONF0_SPEC>;

Register CH%sCONF0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn div_cnt(&self) -> DIV_CNT_R

Bits 0:7 - This register is used to configure the frequency divider’s factor in channel0.

pub fn idle_thres(&self) -> IDLE_THRES_R

Bits 8:23 - In receive mode when no edge is detected on the input signal for longer than reg_idle_thres_ch0 then the receive process is done.

pub fn mem_size(&self) -> MEM_SIZE_R

Bits 24:27 - This register is used to configure the the amount of memory blocks allocated to channel0.

pub fn carrier_en(&self) -> CARRIER_EN_R

Bit 28 - This is the carrier modulation enable control bit for channel0.

pub fn carrier_out_lv(&self) -> CARRIER_OUT_LV_R

Bit 29 - This bit is used to configure the way carrier wave is modulated for channel0.1’b1:transmit on low output level 1’b0:transmit on high output level.

pub fn mem_pd(&self) -> MEM_PD_R

Bit 30 - This bit is used to reduce power consumed by mem. 1:mem is in low power state.

pub fn clk_en(&self) -> CLK_EN_R

Bit 31 - This bit is used to control clock.when software config RMT internal registers it controls the register clock.