Type Alias esp32::rmt::chstatus::R

pub type R = R<CHSTATUS_SPEC>;

Register CH%sSTATUS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn status(&self) -> STATUS_R

Bits 0:31 - The status for channel0

pub fn mem_waddr_ex(&self) -> MEM_WADDR_EX_R

Bits 0:9 - The current memory read address of channel0.

pub fn mem_raddr_ex(&self) -> MEM_RADDR_EX_R

Bits 12:21 - The current memory write address of channel0.

pub fn state(&self) -> STATE_R

Bits 24:26 - The channel0 state machine status register.3’h0 : idle, 3’h1 : send, 3’h2 : read memory, 3’h3 : receive, 3’h4 : wait.

pub fn mem_owner_err(&self) -> MEM_OWNER_ERR_R

Bit 27 - When channel0 is configured for receive mode, this bit will turn to high level if rmt_mem_owner register is not set to 1.

pub fn mem_full(&self) -> MEM_FULL_R

Bit 28 - The memory full status bit for channel0 turns to high level when mem_waddr_ex is greater than or equal to the configuration range.

pub fn mem_empty(&self) -> MEM_EMPTY_R

Bit 29 - The memory empty status bit for channel0. in acyclic mode, this bit turns to high level when mem_raddr_ex is greater than or equal to the configured range.

pub fn apb_mem_wr_err(&self) -> APB_MEM_WR_ERR_R

Bit 30 - The apb write memory status bit for channel0 turns to high level when the apb write address exceeds the configuration range.

pub fn apb_mem_rd_err(&self) -> APB_MEM_RD_ERR_R

Bit 31 - The apb read memory status bit for channel0 turns to high level when the apb read address exceeds the configuration range.