Type Alias esp32::rmt::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ch_tx_end(&self, n: u8) -> CH_TX_END_R

The interrupt raw bit for channel (0-7) turns to high level when the transmit process is done.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_END field

pub fn ch_tx_end_iter(&self) -> impl Iterator<Item = CH_TX_END_R> + '_

Iterator for array of: The interrupt raw bit for channel (0-7) turns to high level when the transmit process is done.

pub fn ch0_tx_end(&self) -> CH_TX_END_R

Bit 0 - The interrupt raw bit for channel 0 turns to high level when the transmit process is done.

pub fn ch1_tx_end(&self) -> CH_TX_END_R

Bit 3 - The interrupt raw bit for channel 1 turns to high level when the transmit process is done.

pub fn ch2_tx_end(&self) -> CH_TX_END_R

Bit 6 - The interrupt raw bit for channel 2 turns to high level when the transmit process is done.

pub fn ch3_tx_end(&self) -> CH_TX_END_R

Bit 9 - The interrupt raw bit for channel 3 turns to high level when the transmit process is done.

pub fn ch4_tx_end(&self) -> CH_TX_END_R

Bit 12 - The interrupt raw bit for channel 4 turns to high level when the transmit process is done.

pub fn ch5_tx_end(&self) -> CH_TX_END_R

Bit 15 - The interrupt raw bit for channel 5 turns to high level when the transmit process is done.

pub fn ch6_tx_end(&self) -> CH_TX_END_R

Bit 18 - The interrupt raw bit for channel 6 turns to high level when the transmit process is done.

pub fn ch7_tx_end(&self) -> CH_TX_END_R

Bit 21 - The interrupt raw bit for channel 7 turns to high level when the transmit process is done.

pub fn ch_rx_end(&self, n: u8) -> CH_RX_END_R

The interrupt raw bit for channel (0-7) turns to high level when the receive process is done.

NOTE: n is number of field in register. n == 0 corresponds to CH0_RX_END field

pub fn ch_rx_end_iter(&self) -> impl Iterator<Item = CH_RX_END_R> + '_

Iterator for array of: The interrupt raw bit for channel (0-7) turns to high level when the receive process is done.

pub fn ch0_rx_end(&self) -> CH_RX_END_R

Bit 1 - The interrupt raw bit for channel 0 turns to high level when the receive process is done.

pub fn ch1_rx_end(&self) -> CH_RX_END_R

Bit 4 - The interrupt raw bit for channel 1 turns to high level when the receive process is done.

pub fn ch2_rx_end(&self) -> CH_RX_END_R

Bit 7 - The interrupt raw bit for channel 2 turns to high level when the receive process is done.

pub fn ch3_rx_end(&self) -> CH_RX_END_R

Bit 10 - The interrupt raw bit for channel 3 turns to high level when the receive process is done.

pub fn ch4_rx_end(&self) -> CH_RX_END_R

Bit 13 - The interrupt raw bit for channel 4 turns to high level when the receive process is done.

pub fn ch5_rx_end(&self) -> CH_RX_END_R

Bit 16 - The interrupt raw bit for channel 5 turns to high level when the receive process is done.

pub fn ch6_rx_end(&self) -> CH_RX_END_R

Bit 19 - The interrupt raw bit for channel 6 turns to high level when the receive process is done.

pub fn ch7_rx_end(&self) -> CH_RX_END_R

Bit 22 - The interrupt raw bit for channel 7 turns to high level when the receive process is done.

pub fn ch_err(&self, n: u8) -> CH_ERR_R

The interrupt raw bit for channel (0-7) turns to high level when channle (0-7) detects some errors.

NOTE: n is number of field in register. n == 0 corresponds to CH0_ERR field

pub fn ch_err_iter(&self) -> impl Iterator<Item = CH_ERR_R> + '_

Iterator for array of: The interrupt raw bit for channel (0-7) turns to high level when channle (0-7) detects some errors.

pub fn ch0_err(&self) -> CH_ERR_R

Bit 2 - The interrupt raw bit for channel 0 turns to high level when channle 0 detects some errors.

pub fn ch1_err(&self) -> CH_ERR_R

Bit 5 - The interrupt raw bit for channel 1 turns to high level when channle 1 detects some errors.

pub fn ch2_err(&self) -> CH_ERR_R

Bit 8 - The interrupt raw bit for channel 2 turns to high level when channle 2 detects some errors.

pub fn ch3_err(&self) -> CH_ERR_R

Bit 11 - The interrupt raw bit for channel 3 turns to high level when channle 3 detects some errors.

pub fn ch4_err(&self) -> CH_ERR_R

Bit 14 - The interrupt raw bit for channel 4 turns to high level when channle 4 detects some errors.

pub fn ch5_err(&self) -> CH_ERR_R

Bit 17 - The interrupt raw bit for channel 5 turns to high level when channle 5 detects some errors.

pub fn ch6_err(&self) -> CH_ERR_R

Bit 20 - The interrupt raw bit for channel 6 turns to high level when channle 6 detects some errors.

pub fn ch7_err(&self) -> CH_ERR_R

Bit 23 - The interrupt raw bit for channel 7 turns to high level when channle 7 detects some errors.

pub fn ch_tx_thr_event(&self, n: u8) -> CH_TX_THR_EVENT_R

The interrupt raw bit for channel (0-7) turns to high level when transmitter in channle(0-7) have send datas more than reg_rmt_tx_lim_ch(0-7) after detecting this interrupt software can updata the old datas with new datas.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_THR_EVENT field

pub fn ch_tx_thr_event_iter( &self ) -> impl Iterator<Item = CH_TX_THR_EVENT_R> + '_

Iterator for array of: The interrupt raw bit for channel (0-7) turns to high level when transmitter in channle(0-7) have send datas more than reg_rmt_tx_lim_ch(0-7) after detecting this interrupt software can updata the old datas with new datas.

pub fn ch0_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 24 - The interrupt raw bit for channel 0 turns to high level when transmitter in channle0 have send datas more than reg_rmt_tx_lim_ch0 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch1_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 25 - The interrupt raw bit for channel 1 turns to high level when transmitter in channle1 have send datas more than reg_rmt_tx_lim_ch1 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch2_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 26 - The interrupt raw bit for channel 2 turns to high level when transmitter in channle2 have send datas more than reg_rmt_tx_lim_ch2 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch3_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 27 - The interrupt raw bit for channel 3 turns to high level when transmitter in channle3 have send datas more than reg_rmt_tx_lim_ch3 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch4_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 28 - The interrupt raw bit for channel 4 turns to high level when transmitter in channle4 have send datas more than reg_rmt_tx_lim_ch4 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch5_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 29 - The interrupt raw bit for channel 5 turns to high level when transmitter in channle5 have send datas more than reg_rmt_tx_lim_ch5 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch6_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 30 - The interrupt raw bit for channel 6 turns to high level when transmitter in channle6 have send datas more than reg_rmt_tx_lim_ch6 after detecting this interrupt software can updata the old datas with new datas.

pub fn ch7_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 31 - The interrupt raw bit for channel 7 turns to high level when transmitter in channle7 have send datas more than reg_rmt_tx_lim_ch7 after detecting this interrupt software can updata the old datas with new datas.