Module esp32::rtc_cntl::reg

Structs

REG_SPEC

You can read this register and get reg::R. You can reset, write, write_with_zero this register using reg::W. You can also modify this register. See API.

Type Aliases

DBIAS_SLP_R

Field DBIAS_SLP reader - RTC_DBIAS during sleep

DBIAS_SLP_W

Field DBIAS_SLP writer - RTC_DBIAS during sleep

DBIAS_WAK_R

Field DBIAS_WAK reader - RTC_DBIAS during wakeup

DBIAS_WAK_W

Field DBIAS_WAK writer - RTC_DBIAS during wakeup

DBOOST_FORCE_PD_R

Field DBOOST_FORCE_PD reader - RTC_DBOOST force power down

DBOOST_FORCE_PD_W

Field DBOOST_FORCE_PD writer - RTC_DBOOST force power down

DBOOST_FORCE_PU_R

Field DBOOST_FORCE_PU reader - RTC_DBOOST force power up

DBOOST_FORCE_PU_W

Field DBOOST_FORCE_PU writer - RTC_DBOOST force power up

DIG_DBIAS_SLP_R

Field DIG_DBIAS_SLP reader - DIG_REG_DBIAS during sleep

DIG_DBIAS_SLP_W

Field DIG_DBIAS_SLP writer - DIG_REG_DBIAS during sleep

DIG_DBIAS_WAK_R

Field DIG_DBIAS_WAK reader - DIG_REG_DBIAS during wakeup

DIG_DBIAS_WAK_W

Field DIG_DBIAS_WAK writer - DIG_REG_DBIAS during wakeup

FORCE_PD_R

Field FORCE_PD reader - RTC_REG force power down (for RTC_REG power down means decrease the voltage to 0.8v or lower )

FORCE_PD_W

Field FORCE_PD writer - RTC_REG force power down (for RTC_REG power down means decrease the voltage to 0.8v or lower )

FORCE_PU_R

Field FORCE_PU reader - RTC_REG force power up

FORCE_PU_W

Field FORCE_PU writer - RTC_REG force power up

R

Register REG reader

SCK_DCAP_FORCE_R

Field SCK_DCAP_FORCE reader - N/A

SCK_DCAP_FORCE_W

Field SCK_DCAP_FORCE writer - N/A

SCK_DCAP_R

Field SCK_DCAP reader - SCK_DCAP

SCK_DCAP_W

Field SCK_DCAP writer - SCK_DCAP

W

Register REG writer