Module esp32::rtc_i2c::debug_status

Structs

DEBUG_STATUS_SPEC

You can read this register and get debug_status::R. You can reset, write, write_with_zero this register using debug_status::W. You can also modify this register. See API.

Type Aliases

ACK_VAL_R

Field ACK_VAL reader - The value of an acknowledge signal on the bus

ACK_VAL_W

Field ACK_VAL writer - The value of an acknowledge signal on the bus

ARB_LOST_R

Field ARB_LOST reader - When working as a master, lost control of I2C bus

ARB_LOST_W

Field ARB_LOST writer - When working as a master, lost control of I2C bus

BUS_BUSY_R

Field BUS_BUSY reader - operation is in progress

BUS_BUSY_W

Field BUS_BUSY writer - operation is in progress

BYTE_TRANS_R

Field BYTE_TRANS reader - 8 bit transmit done

BYTE_TRANS_W

Field BYTE_TRANS writer - 8 bit transmit done

MAIN_STATE_R

Field MAIN_STATE reader - state of the main state machine

MAIN_STATE_W

Field MAIN_STATE writer - state of the main state machine

R

Register DEBUG_STATUS reader

SCL_STATE_R

Field SCL_STATE reader - state of SCL state machine

SCL_STATE_W

Field SCL_STATE writer - state of SCL state machine

SLAVE_ADDR_MATCH_R

Field SLAVE_ADDR_MATCH reader - When working as a slave, whether address was matched

SLAVE_ADDR_MATCH_W

Field SLAVE_ADDR_MATCH writer - When working as a slave, whether address was matched

SLAVE_RW_R

Field SLAVE_RW reader - When working as a slave, the value of R/W bit received

SLAVE_RW_W

Field SLAVE_RW writer - When working as a slave, the value of R/W bit received

TIMED_OUT_R

Field TIMED_OUT reader - Transfer has timed out

TIMED_OUT_W

Field TIMED_OUT writer - Transfer has timed out

W

Register DEBUG_STATUS writer