Type Alias esp32::rtc_io::adc_pad::R

pub type R = R<ADC_PAD_SPEC>;

Register ADC_PAD reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn adc2_fun_ie(&self) -> ADC2_FUN_IE_R

Bit 18 - the input enable of the pad

pub fn adc2_slp_ie(&self) -> ADC2_SLP_IE_R

Bit 19 - the input enable of the pad in sleep status

pub fn adc2_slp_sel(&self) -> ADC2_SLP_SEL_R

Bit 20 - the sleep status selection signal of the pad

pub fn adc2_fun_sel(&self) -> ADC2_FUN_SEL_R

Bits 21:22 - the functional selection signal of the pad

pub fn adc1_fun_ie(&self) -> ADC1_FUN_IE_R

Bit 23 - the input enable of the pad

pub fn adc1_slp_ie(&self) -> ADC1_SLP_IE_R

Bit 24 - the input enable of the pad in sleep status

pub fn adc1_slp_sel(&self) -> ADC1_SLP_SEL_R

Bit 25 - the sleep status selection signal of the pad

pub fn adc1_fun_sel(&self) -> ADC1_FUN_SEL_R

Bits 26:27 - the functional selection signal of the pad

pub fn adc2_mux_sel(&self) -> ADC2_MUX_SEL_R

Bit 28 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn adc1_mux_sel(&self) -> ADC1_MUX_SEL_R

Bit 29 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn adc2_hold(&self) -> ADC2_HOLD_R

Bit 30 - hold the current value of the output when setting the hold to Ò1Ó

pub fn adc1_hold(&self) -> ADC1_HOLD_R

Bit 31 - hold the current value of the output when setting the hold to Ò1Ó