Type Alias esp32::rtc_io::adc_pad::W

pub type W = W<ADC_PAD_SPEC>;

Register ADC_PAD writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn adc2_fun_ie(&mut self) -> ADC2_FUN_IE_W<'_, ADC_PAD_SPEC>

Bit 18 - the input enable of the pad

pub fn adc2_slp_ie(&mut self) -> ADC2_SLP_IE_W<'_, ADC_PAD_SPEC>

Bit 19 - the input enable of the pad in sleep status

pub fn adc2_slp_sel(&mut self) -> ADC2_SLP_SEL_W<'_, ADC_PAD_SPEC>

Bit 20 - the sleep status selection signal of the pad

pub fn adc2_fun_sel(&mut self) -> ADC2_FUN_SEL_W<'_, ADC_PAD_SPEC>

Bits 21:22 - the functional selection signal of the pad

pub fn adc1_fun_ie(&mut self) -> ADC1_FUN_IE_W<'_, ADC_PAD_SPEC>

Bit 23 - the input enable of the pad

pub fn adc1_slp_ie(&mut self) -> ADC1_SLP_IE_W<'_, ADC_PAD_SPEC>

Bit 24 - the input enable of the pad in sleep status

pub fn adc1_slp_sel(&mut self) -> ADC1_SLP_SEL_W<'_, ADC_PAD_SPEC>

Bit 25 - the sleep status selection signal of the pad

pub fn adc1_fun_sel(&mut self) -> ADC1_FUN_SEL_W<'_, ADC_PAD_SPEC>

Bits 26:27 - the functional selection signal of the pad

pub fn adc2_mux_sel(&mut self) -> ADC2_MUX_SEL_W<'_, ADC_PAD_SPEC>

Bit 28 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn adc1_mux_sel(&mut self) -> ADC1_MUX_SEL_W<'_, ADC_PAD_SPEC>

Bit 29 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn adc2_hold(&mut self) -> ADC2_HOLD_W<'_, ADC_PAD_SPEC>

Bit 30 - hold the current value of the output when setting the hold to Ò1Ó

pub fn adc1_hold(&mut self) -> ADC1_HOLD_W<'_, ADC_PAD_SPEC>

Bit 31 - hold the current value of the output when setting the hold to Ò1Ó

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual