Type Alias esp32::rtc_io::pad_dac1::W

pub type W = W<PAD_DAC1_SPEC>;

Register PAD_DAC1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn pdac1_dac_xpd_force( &mut self ) -> PDAC1_DAC_XPD_FORCE_W<'_, PAD_DAC1_SPEC>

Bit 10 - Power on DAC1. Usually we need to tristate PDAC1 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

pub fn pdac1_fun_ie(&mut self) -> PDAC1_FUN_IE_W<'_, PAD_DAC1_SPEC>

Bit 11 - the input enable of the pad

pub fn pdac1_slp_oe(&mut self) -> PDAC1_SLP_OE_W<'_, PAD_DAC1_SPEC>

Bit 12 - the output enable of the pad in sleep status

pub fn pdac1_slp_ie(&mut self) -> PDAC1_SLP_IE_W<'_, PAD_DAC1_SPEC>

Bit 13 - the input enable of the pad in sleep status

pub fn pdac1_slp_sel(&mut self) -> PDAC1_SLP_SEL_W<'_, PAD_DAC1_SPEC>

Bit 14 - the sleep status selection signal of the pad

pub fn pdac1_fun_sel(&mut self) -> PDAC1_FUN_SEL_W<'_, PAD_DAC1_SPEC>

Bits 15:16 - the functional selection signal of the pad

pub fn pdac1_mux_sel(&mut self) -> PDAC1_MUX_SEL_W<'_, PAD_DAC1_SPEC>

Bit 17 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn pdac1_xpd_dac(&mut self) -> PDAC1_XPD_DAC_W<'_, PAD_DAC1_SPEC>

Bit 18 - Power on DAC1. Usually we need to tristate PDAC1 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

pub fn pdac1_dac(&mut self) -> PDAC1_DAC_W<'_, PAD_DAC1_SPEC>

Bits 19:26 - PAD DAC1 control code.

pub fn pdac1_rue(&mut self) -> PDAC1_RUE_W<'_, PAD_DAC1_SPEC>

Bit 27 - the pull up enable of the pad

pub fn pdac1_rde(&mut self) -> PDAC1_RDE_W<'_, PAD_DAC1_SPEC>

Bit 28 - the pull down enable of the pad

pub fn pdac1_hold(&mut self) -> PDAC1_HOLD_W<'_, PAD_DAC1_SPEC>

Bit 29 - hold the current value of the output when setting the hold to Ò1Ó

pub fn pdac1_drv(&mut self) -> PDAC1_DRV_W<'_, PAD_DAC1_SPEC>

Bits 30:31 - the driver strength of the pad

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual