Module esp32::rtc_io::pad_dac2

Structs

PAD_DAC2_SPEC

You can read this register and get pad_dac2::R. You can reset, write, write_with_zero this register using pad_dac2::W. You can also modify this register. See API.

Type Aliases

PDAC2_DAC_R

Field PDAC2_DAC reader - PAD DAC2 control code.

PDAC2_DAC_W

Field PDAC2_DAC writer - PAD DAC2 control code.

PDAC2_DAC_XPD_FORCE_R

Field PDAC2_DAC_XPD_FORCE reader - Power on DAC2. Usually we need to tristate PDAC2 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

PDAC2_DAC_XPD_FORCE_W

Field PDAC2_DAC_XPD_FORCE writer - Power on DAC2. Usually we need to tristate PDAC2 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

PDAC2_DRV_R

Field PDAC2_DRV reader - the driver strength of the pad

PDAC2_DRV_W

Field PDAC2_DRV writer - the driver strength of the pad

PDAC2_FUN_IE_R

Field PDAC2_FUN_IE reader - the input enable of the pad

PDAC2_FUN_IE_W

Field PDAC2_FUN_IE writer - the input enable of the pad

PDAC2_FUN_SEL_R

Field PDAC2_FUN_SEL reader - the functional selection signal of the pad

PDAC2_FUN_SEL_W

Field PDAC2_FUN_SEL writer - the functional selection signal of the pad

PDAC2_HOLD_R

Field PDAC2_HOLD reader - hold the current value of the output when setting the hold to Ò1Ó

PDAC2_HOLD_W

Field PDAC2_HOLD writer - hold the current value of the output when setting the hold to Ò1Ó

PDAC2_MUX_SEL_R

Field PDAC2_MUX_SEL reader - Ò1Ó select the digital function Ó0Óslection the rtc function

PDAC2_MUX_SEL_W

Field PDAC2_MUX_SEL writer - Ò1Ó select the digital function Ó0Óslection the rtc function

PDAC2_RDE_R

Field PDAC2_RDE reader - the pull down enable of the pad

PDAC2_RDE_W

Field PDAC2_RDE writer - the pull down enable of the pad

PDAC2_RUE_R

Field PDAC2_RUE reader - the pull up enable of the pad

PDAC2_RUE_W

Field PDAC2_RUE writer - the pull up enable of the pad

PDAC2_SLP_IE_R

Field PDAC2_SLP_IE reader - the input enable of the pad in sleep status

PDAC2_SLP_IE_W

Field PDAC2_SLP_IE writer - the input enable of the pad in sleep status

PDAC2_SLP_OE_R

Field PDAC2_SLP_OE reader - the output enable of the pad in sleep status

PDAC2_SLP_OE_W

Field PDAC2_SLP_OE writer - the output enable of the pad in sleep status

PDAC2_SLP_SEL_R

Field PDAC2_SLP_SEL reader - the sleep status selection signal of the pad

PDAC2_SLP_SEL_W

Field PDAC2_SLP_SEL writer - the sleep status selection signal of the pad

PDAC2_XPD_DAC_R

Field PDAC2_XPD_DAC reader - Power on DAC2. Usually we need to tristate PDAC1 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

PDAC2_XPD_DAC_W

Field PDAC2_XPD_DAC writer - Power on DAC2. Usually we need to tristate PDAC1 if we power on the DAC i.e. IE=0 OE=0 RDE=0 RUE=0

R

Register PAD_DAC2 reader

W

Register PAD_DAC2 writer