Module esp32::rtc_io::touch_pad1

Structs

TOUCH_PAD1_SPEC

You can read this register and get touch_pad1::R. You can reset, write, write_with_zero this register using touch_pad1::W. You can also modify this register. See API.

Type Aliases

DAC_R

Field DAC reader - touch sensor slope control. 3-bit for each touch panel default 100.

DAC_W

Field DAC writer - touch sensor slope control. 3-bit for each touch panel default 100.

DRV_R

Field DRV reader - the driver strength of the pad

DRV_W

Field DRV writer - the driver strength of the pad

FUN_IE_R

Field FUN_IE reader - the input enable of the pad

FUN_IE_W

Field FUN_IE writer - the input enable of the pad

FUN_SEL_R

Field FUN_SEL reader - the functional selection signal of the pad

FUN_SEL_W

Field FUN_SEL writer - the functional selection signal of the pad

HOLD_R

Field HOLD reader -

HOLD_W

Field HOLD writer -

MUX_SEL_R

Field MUX_SEL reader - Ò1Ó select the digital function Ó0Óslection the rtc function

MUX_SEL_W

Field MUX_SEL writer - Ò1Ó select the digital function Ó0Óslection the rtc function

R

Register TOUCH_PAD1 reader

RDE_R

Field RDE reader - the pull down enable of the pad

RDE_W

Field RDE writer - the pull down enable of the pad

RUE_R

Field RUE reader - the pull up enable of the pad

RUE_W

Field RUE writer - the pull up enable of the pad

SLP_IE_R

Field SLP_IE reader - the input enable of the pad in sleep status

SLP_IE_W

Field SLP_IE writer - the input enable of the pad in sleep status

SLP_OE_R

Field SLP_OE reader - the output enable of the pad in sleep status

SLP_OE_W

Field SLP_OE writer - the output enable of the pad in sleep status

SLP_SEL_R

Field SLP_SEL reader - the sleep status selection signal of the pad

SLP_SEL_W

Field SLP_SEL writer - the sleep status selection signal of the pad

START_R

Field START reader - start touch sensor.

START_W

Field START writer - start touch sensor.

TIE_OPT_R

Field TIE_OPT reader - default touch sensor tie option. 0: tie low 1: tie high.

TIE_OPT_W

Field TIE_OPT writer - default touch sensor tie option. 0: tie low 1: tie high.

TO_GPIO_R

Field TO_GPIO reader - connect the rtc pad input to digital pad input Ó0Ó is availbale.GPIO0

TO_GPIO_W

Field TO_GPIO writer - connect the rtc pad input to digital pad input Ó0Ó is availbale.GPIO0

W

Register TOUCH_PAD1 writer

XPD_R

Field XPD reader - touch sensor power on.

XPD_W

Field XPD writer - touch sensor power on.