Type Alias esp32::rtc_io::xtal_32k_pad::R

pub type R = R<XTAL_32K_PAD_SPEC>;

Register XTAL_32K_PAD reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn dbias_xtal_32k(&self) -> DBIAS_XTAL_32K_R

Bits 1:2 - 32K XTAL self-bias reference control.

pub fn dres_xtal_32k(&self) -> DRES_XTAL_32K_R

Bits 3:4 - 32K XTAL resistor bias control.

pub fn x32p_fun_ie(&self) -> X32P_FUN_IE_R

Bit 5 - the input enable of the pad

pub fn x32p_slp_oe(&self) -> X32P_SLP_OE_R

Bit 6 - the output enable of the pad in sleep status

pub fn x32p_slp_ie(&self) -> X32P_SLP_IE_R

Bit 7 - the input enable of the pad in sleep status

pub fn x32p_slp_sel(&self) -> X32P_SLP_SEL_R

Bit 8 - the sleep status selection signal of the pad

pub fn x32p_fun_sel(&self) -> X32P_FUN_SEL_R

Bits 9:10 - the functional selection signal of the pad

pub fn x32n_fun_ie(&self) -> X32N_FUN_IE_R

Bit 11 - the input enable of the pad

pub fn x32n_slp_oe(&self) -> X32N_SLP_OE_R

Bit 12 - the output enable of the pad in sleep status

pub fn x32n_slp_ie(&self) -> X32N_SLP_IE_R

Bit 13 - the input enable of the pad in sleep status

pub fn x32n_slp_sel(&self) -> X32N_SLP_SEL_R

Bit 14 - the sleep status selection signal of the pad

pub fn x32n_fun_sel(&self) -> X32N_FUN_SEL_R

Bits 15:16 - the functional selection signal of the pad

pub fn x32p_mux_sel(&self) -> X32P_MUX_SEL_R

Bit 17 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn x32n_mux_sel(&self) -> X32N_MUX_SEL_R

Bit 18 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn xpd_xtal_32k(&self) -> XPD_XTAL_32K_R

Bit 19 - Power up 32kHz crystal oscillator

pub fn dac_xtal_32k(&self) -> DAC_XTAL_32K_R

Bits 20:21 - 32K XTAL bias current DAC.

pub fn x32p_rue(&self) -> X32P_RUE_R

Bit 22 - the pull up enable of the pad

pub fn x32p_rde(&self) -> X32P_RDE_R

Bit 23 - the pull down enable of the pad

pub fn x32p_hold(&self) -> X32P_HOLD_R

Bit 24 - hold the current value of the output when setting the hold to Ò1Ó

pub fn x32p_drv(&self) -> X32P_DRV_R

Bits 25:26 - the driver strength of the pad

pub fn x32n_rue(&self) -> X32N_RUE_R

Bit 27 - the pull up enable of the pad

pub fn x32n_rde(&self) -> X32N_RDE_R

Bit 28 - the pull down enable of the pad

pub fn x32n_hold(&self) -> X32N_HOLD_R

Bit 29 - hold the current value of the output when setting the hold to Ò1Ó

pub fn x32n_drv(&self) -> X32N_DRV_R

Bits 30:31 - the driver strength of the pad