Type Alias esp32::rtc_io::xtal_32k_pad::W

pub type W = W<XTAL_32K_PAD_SPEC>;

Register XTAL_32K_PAD writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn dbias_xtal_32k(&mut self) -> DBIAS_XTAL_32K_W<'_, XTAL_32K_PAD_SPEC>

Bits 1:2 - 32K XTAL self-bias reference control.

pub fn dres_xtal_32k(&mut self) -> DRES_XTAL_32K_W<'_, XTAL_32K_PAD_SPEC>

Bits 3:4 - 32K XTAL resistor bias control.

pub fn x32p_fun_ie(&mut self) -> X32P_FUN_IE_W<'_, XTAL_32K_PAD_SPEC>

Bit 5 - the input enable of the pad

pub fn x32p_slp_oe(&mut self) -> X32P_SLP_OE_W<'_, XTAL_32K_PAD_SPEC>

Bit 6 - the output enable of the pad in sleep status

pub fn x32p_slp_ie(&mut self) -> X32P_SLP_IE_W<'_, XTAL_32K_PAD_SPEC>

Bit 7 - the input enable of the pad in sleep status

pub fn x32p_slp_sel(&mut self) -> X32P_SLP_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bit 8 - the sleep status selection signal of the pad

pub fn x32p_fun_sel(&mut self) -> X32P_FUN_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bits 9:10 - the functional selection signal of the pad

pub fn x32n_fun_ie(&mut self) -> X32N_FUN_IE_W<'_, XTAL_32K_PAD_SPEC>

Bit 11 - the input enable of the pad

pub fn x32n_slp_oe(&mut self) -> X32N_SLP_OE_W<'_, XTAL_32K_PAD_SPEC>

Bit 12 - the output enable of the pad in sleep status

pub fn x32n_slp_ie(&mut self) -> X32N_SLP_IE_W<'_, XTAL_32K_PAD_SPEC>

Bit 13 - the input enable of the pad in sleep status

pub fn x32n_slp_sel(&mut self) -> X32N_SLP_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bit 14 - the sleep status selection signal of the pad

pub fn x32n_fun_sel(&mut self) -> X32N_FUN_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bits 15:16 - the functional selection signal of the pad

pub fn x32p_mux_sel(&mut self) -> X32P_MUX_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bit 17 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn x32n_mux_sel(&mut self) -> X32N_MUX_SEL_W<'_, XTAL_32K_PAD_SPEC>

Bit 18 - Ò1Ó select the digital function Ó0Óslection the rtc function

pub fn xpd_xtal_32k(&mut self) -> XPD_XTAL_32K_W<'_, XTAL_32K_PAD_SPEC>

Bit 19 - Power up 32kHz crystal oscillator

pub fn dac_xtal_32k(&mut self) -> DAC_XTAL_32K_W<'_, XTAL_32K_PAD_SPEC>

Bits 20:21 - 32K XTAL bias current DAC.

pub fn x32p_rue(&mut self) -> X32P_RUE_W<'_, XTAL_32K_PAD_SPEC>

Bit 22 - the pull up enable of the pad

pub fn x32p_rde(&mut self) -> X32P_RDE_W<'_, XTAL_32K_PAD_SPEC>

Bit 23 - the pull down enable of the pad

pub fn x32p_hold(&mut self) -> X32P_HOLD_W<'_, XTAL_32K_PAD_SPEC>

Bit 24 - hold the current value of the output when setting the hold to Ò1Ó

pub fn x32p_drv(&mut self) -> X32P_DRV_W<'_, XTAL_32K_PAD_SPEC>

Bits 25:26 - the driver strength of the pad

pub fn x32n_rue(&mut self) -> X32N_RUE_W<'_, XTAL_32K_PAD_SPEC>

Bit 27 - the pull up enable of the pad

pub fn x32n_rde(&mut self) -> X32N_RDE_W<'_, XTAL_32K_PAD_SPEC>

Bit 28 - the pull down enable of the pad

pub fn x32n_hold(&mut self) -> X32N_HOLD_W<'_, XTAL_32K_PAD_SPEC>

Bit 29 - hold the current value of the output when setting the hold to Ò1Ó

pub fn x32n_drv(&mut self) -> X32N_DRV_W<'_, XTAL_32K_PAD_SPEC>

Bits 30:31 - the driver strength of the pad

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual