Type Alias esp32::sdhost::clk_edge_sel::R

pub type R = R<CLK_EDGE_SEL_SPEC>;

Register CLK_EDGE_SEL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn cclkin_edge_drv_sel(&self) -> CCLKIN_EDGE_DRV_SEL_R

Bits 0:2 - It’s used to select the clock phase of the output signal from phase 0, phase 90, phase 180, phase 270.

pub fn cclkin_edge_sam_sel(&self) -> CCLKIN_EDGE_SAM_SEL_R

Bits 3:5 - It’s used to select the clock phase of the input signal from phase 0, phase 90, phase 180, phase 270.

pub fn cclkin_edge_slf_sel(&self) -> CCLKIN_EDGE_SLF_SEL_R

Bits 6:8 - It’s used to select the clock phase of the internal signal from phase 0, phase 90, phase 180, phase 270.

pub fn ccllkin_edge_h(&self) -> CCLLKIN_EDGE_H_R

Bits 9:12 - The high level of the divider clock. The value should be smaller than CCLKIN_EDGE_L.

pub fn ccllkin_edge_l(&self) -> CCLLKIN_EDGE_L_R

Bits 13:16 - The low level of the divider clock. The value should be larger than CCLKIN_EDGE_H.

pub fn ccllkin_edge_n(&self) -> CCLLKIN_EDGE_N_R

Bits 17:20 - The value should be equal to CCLKIN_EDGE_L.

pub fn esdio_mode(&self) -> ESDIO_MODE_R

Bit 21 - Enable esdio mode.

pub fn esd_mode(&self) -> ESD_MODE_R

Bit 22 - Enable esd mode.

pub fn cclk_en(&self) -> CCLK_EN_R

Bit 23 - Sdio clock enable