Type Alias esp32::sdhost::cmd::UPDATE_CLOCK_REGISTERS_ONLY_R

pub type UPDATE_CLOCK_REGISTERS_ONLY_R = BitReader;

Field UPDATE_CLOCK_REGISTERS_ONLY reader - 0: Normal command sequence; 1: Do not send commands, just update clock register value into card clock domain. Following register values are transferred into card clock domain: CLKDIV, CLRSRC, and CLKENA. Changes card clocks (change frequency, truncate off or on, and set low-frequency mode). This is provided in order to change clock frequency or stop clock without having to send command to cards. During normal command sequence, when sdhost_update_clock_registers_only = 0, following control registers are transferred from BIU to CIU: CMD, CMDARG, TMOUT, CTYPE, BLKSIZ, and BYTCNT. CIU uses new register values for new command sequence to card(s). When bit is set, there are no Command Done interrupts because no command is sent to SD_MMC_CEATA cards.

Aliased Type

struct UPDATE_CLOCK_REGISTERS_ONLY_R { /* private fields */ }