Type Alias esp32::sdhost::status::R

pub type R = R<STATUS_SPEC>;

Register STATUS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn fifo_rx_watermark(&self) -> FIFO_RX_WATERMARK_R

Bit 0 - FIFO reached Receive watermark level, not qualified with data transfer.

pub fn fifo_tx_watermark(&self) -> FIFO_TX_WATERMARK_R

Bit 1 - FIFO reached Transmit watermark level, not qualified with data transfer.

pub fn fifo_empty(&self) -> FIFO_EMPTY_R

Bit 2 - FIFO is empty status.

pub fn fifo_full(&self) -> FIFO_FULL_R

Bit 3 - FIFO is full status.

pub fn command_fsm_states(&self) -> COMMAND_FSM_STATES_R

Bits 4:7 - Command FSM states. 0: Idle; 1: Send init sequence; 2: Send cmd start bit; 3: Send cmd tx bit; 4: Send cmd index + arg; 5: Send cmd crc7; 6: Send cmd end bit; 7: Receive resp start bit; 8: Receive resp IRQ response; 9: Receive resp tx bit; 10: Receive resp cmd idx; 11: Receive resp data; 12: Receive resp crc7; 13: Receive resp end bit; 14: Cmd path wait NCC; 15: Wait, cmd-to-response turnaround.

pub fn data_3_status(&self) -> DATA_3_STATUS_R

Bit 8 - Raw selected sdhost_card_data[3], checks whether card is present. 0: card not present; 1: card present.

pub fn data_busy(&self) -> DATA_BUSY_R

Bit 9 - Inverted version of raw selected sdhost_card_data[0]. 0: Card data not busy; 1: Card data busy.

pub fn data_state_mc_busy(&self) -> DATA_STATE_MC_BUSY_R

Bit 10 - Data transmit or receive state-machine is busy.

pub fn response_index(&self) -> RESPONSE_INDEX_R

Bits 11:16 - Index of previous response, including any auto-stop sent by core.

pub fn fifo_count(&self) -> FIFO_COUNT_R

Bits 17:29 - FIFO count, number of filled locations in FIFO.