Module esp32::spi0::cmd

Structs

CMD_SPEC

You can read this register and get cmd::R. You can reset, write, write_with_zero this register using cmd::W. You can also modify this register. See API.

Type Aliases

FLASH_BE_R

Field FLASH_BE reader - Block erase enable. A 64KB block is erased via SPI command D8H. Block erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_BE_W

Field FLASH_BE writer - Block erase enable. A 64KB block is erased via SPI command D8H. Block erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_CE_R

Field FLASH_CE reader - Chip erase enable. Chip erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_CE_W

Field FLASH_CE writer - Chip erase enable. Chip erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_DP_R

Field FLASH_DP reader - Drive Flash into power down. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_DP_W

Field FLASH_DP writer - Drive Flash into power down. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_HPM_R

Field FLASH_HPM reader - Drive Flash into high performance mode. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_HPM_W

Field FLASH_HPM writer - Drive Flash into high performance mode. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_PER_R

Field FLASH_PER reader - program erase resume bit program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_PER_W

Field FLASH_PER writer - program erase resume bit program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_PES_R

Field FLASH_PES reader - program erase suspend bit program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_PES_W

Field FLASH_PES writer - program erase suspend bit program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_PP_R

Field FLASH_PP reader - Page program enable(1 byte ~256 bytes data to be programmed). Page program operation will be triggered when the bit is set. The bit will be cleared once the operation done .1: enable 0: disable.

FLASH_PP_W

Field FLASH_PP writer - Page program enable(1 byte ~256 bytes data to be programmed). Page program operation will be triggered when the bit is set. The bit will be cleared once the operation done .1: enable 0: disable.

FLASH_RDID_R

Field FLASH_RDID reader - Read JEDEC ID . Read ID command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_RDID_W

Field FLASH_RDID writer - Read JEDEC ID . Read ID command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_RDSR_R

Field FLASH_RDSR reader - Read status register-1. Read status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_RDSR_W

Field FLASH_RDSR writer - Read status register-1. Read status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_READ_R

Field FLASH_READ reader - Read flash enable. Read flash operation will be triggered when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_READ_W

Field FLASH_READ writer - Read flash enable. Read flash operation will be triggered when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_RES_R

Field FLASH_RES reader - This bit combined with reg_resandres bit releases Flash from the power-down state or high performance mode and obtains the devices ID. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_RES_W

Field FLASH_RES writer - This bit combined with reg_resandres bit releases Flash from the power-down state or high performance mode and obtains the devices ID. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_SE_R

Field FLASH_SE reader - Sector erase enable. A 4KB sector is erased via SPI command 20H. Sector erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_SE_W

Field FLASH_SE writer - Sector erase enable. A 4KB sector is erased via SPI command 20H. Sector erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_WRDI_R

Field FLASH_WRDI reader - Write flash disable. Write disable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_WRDI_W

Field FLASH_WRDI writer - Write flash disable. Write disable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_WREN_R

Field FLASH_WREN reader - Write flash enable. Write enable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_WREN_W

Field FLASH_WREN writer - Write flash enable. Write enable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

FLASH_WRSR_R

Field FLASH_WRSR reader - Write status register enable. Write status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

FLASH_WRSR_W

Field FLASH_WRSR writer - Write status register enable. Write status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

R

Register CMD reader

USR_R

Field USR reader - User define command enable. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

USR_W

Field USR writer - User define command enable. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

W

Register CMD writer