Module esp32::spi0::ctrl

Structs

CTRL_SPEC

You can read this register and get ctrl::R. You can reset, write, write_with_zero this register using ctrl::W. You can also modify this register. See API.

Type Aliases

FASTRD_MODE_R

Field FASTRD_MODE reader - This bit enable the bits: spi_fread_qio spi_fread_dio spi_fread_qout and spi_fread_dout. 1: enable 0: disable.

FASTRD_MODE_W

Field FASTRD_MODE writer - This bit enable the bits: spi_fread_qio spi_fread_dio spi_fread_qout and spi_fread_dout. 1: enable 0: disable.

FCS_CRC_EN_R

Field FCS_CRC_EN reader - For SPI1 initialize crc32 module before writing encrypted data to flash. Active low.

FCS_CRC_EN_W

Field FCS_CRC_EN writer - For SPI1 initialize crc32 module before writing encrypted data to flash. Active low.

FREAD_DIO_R

Field FREAD_DIO reader - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DIO_W

Field FREAD_DIO writer - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DUAL_R

Field FREAD_DUAL reader - In the read operations read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_DUAL_W

Field FREAD_DUAL writer - In the read operations read-data phase apply 2 signals. 1: enable 0: disable.

FREAD_QIO_R

Field FREAD_QIO reader - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QIO_W

Field FREAD_QIO writer - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QUAD_R

Field FREAD_QUAD reader - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

FREAD_QUAD_W

Field FREAD_QUAD writer - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

R

Register CTRL reader

RD_BIT_ORDER_R

Field RD_BIT_ORDER reader - In read-data (MISO) phase 1: LSB first 0: MSB first

RD_BIT_ORDER_W

Field RD_BIT_ORDER writer - In read-data (MISO) phase 1: LSB first 0: MSB first

RESANDRES_R

Field RESANDRES reader - The Device ID is read out to SPI_RD_STATUS register, this bit combine with spi_flash_res bit. 1: enable 0: disable.

RESANDRES_W

Field RESANDRES writer - The Device ID is read out to SPI_RD_STATUS register, this bit combine with spi_flash_res bit. 1: enable 0: disable.

TX_CRC_EN_R

Field TX_CRC_EN reader - For SPI1 enable crc32 when writing encrypted data to flash. 1: enable 0:disable

TX_CRC_EN_W

Field TX_CRC_EN writer - For SPI1 enable crc32 when writing encrypted data to flash. 1: enable 0:disable

W

Register CTRL writer

WAIT_FLASH_IDLE_EN_R

Field WAIT_FLASH_IDLE_EN reader - wait flash idle when program flash or erase flash. 1: enable 0: disable.

WAIT_FLASH_IDLE_EN_W

Field WAIT_FLASH_IDLE_EN writer - wait flash idle when program flash or erase flash. 1: enable 0: disable.

WP_R

Field WP reader - Write protect signal output when SPI is idle. 1: output high 0: output low.

WP_W

Field WP writer - Write protect signal output when SPI is idle. 1: output high 0: output low.

WRSR_2B_R

Field WRSR_2B reader - two bytes data will be written to status register when it is set. 1: enable 0: disable.

WRSR_2B_W

Field WRSR_2B writer - two bytes data will be written to status register when it is set. 1: enable 0: disable.

WR_BIT_ORDER_R

Field WR_BIT_ORDER reader - In command address write-data (MOSI) phases 1: LSB firs 0: MSB first

WR_BIT_ORDER_W

Field WR_BIT_ORDER writer - In command address write-data (MOSI) phases 1: LSB firs 0: MSB first