Type Alias esp32::spi0::dma_int_clr::W

pub type W = W<DMA_INT_CLR_SPEC>;

Register DMA_INT_CLR writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn inlink_dscr_empty_int_clr( &mut self ) -> INLINK_DSCR_EMPTY_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 0 - The clear bit for lack of enough inlink descriptors.

pub fn outlink_dscr_error_int_clr( &mut self ) -> OUTLINK_DSCR_ERROR_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 1 - The clear bit for outlink descriptor error.

pub fn inlink_dscr_error_int_clr( &mut self ) -> INLINK_DSCR_ERROR_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 2 - The clear bit for inlink descriptor error.

pub fn in_done_int_clr(&mut self) -> IN_DONE_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 3 - The clear bit for completing usage of a inlink descriptor.

pub fn in_err_eof_int_clr( &mut self ) -> IN_ERR_EOF_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 4 - The clear bit for receiving error.

pub fn in_suc_eof_int_clr( &mut self ) -> IN_SUC_EOF_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 5 - The clear bit for completing receiving all the packets from host.

pub fn out_done_int_clr(&mut self) -> OUT_DONE_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 6 - The clear bit for completing usage of a outlink descriptor.

pub fn out_eof_int_clr(&mut self) -> OUT_EOF_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 7 - The clear bit for sending a packet to host done.

pub fn out_total_eof_int_clr( &mut self ) -> OUT_TOTAL_EOF_INT_CLR_W<'_, DMA_INT_CLR_SPEC>

Bit 8 - The clear bit for sending all the packets to host done.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual