Module esp32::spi0::pin

Structs

PIN_SPEC

You can read this register and get pin::R. You can reset, write, write_with_zero this register using pin::W. You can also modify this register. See API.

Type Aliases

CK_DIS_R

Field CK_DIS reader - 1: spi clk out disable 0: spi clk out enable

CK_DIS_W

Field CK_DIS writer - 1: spi clk out disable 0: spi clk out enable

CK_IDLE_EDGE_R

Field CK_IDLE_EDGE reader - 1: spi clk line is high when idle 0: spi clk line is low when idle

CK_IDLE_EDGE_W

Field CK_IDLE_EDGE writer - 1: spi clk line is high when idle 0: spi clk line is low when idle

CS0_DIS_R

Field CS0_DIS reader - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin

CS0_DIS_W

Field CS0_DIS writer - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin

CS1_DIS_R

Field CS1_DIS reader - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin

CS1_DIS_W

Field CS1_DIS writer - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin

CS2_DIS_R

Field CS2_DIS reader - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin

CS2_DIS_W

Field CS2_DIS writer - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin

CS_KEEP_ACTIVE_R

Field CS_KEEP_ACTIVE reader - spi cs line keep low when the bit is set.

CS_KEEP_ACTIVE_W

Field CS_KEEP_ACTIVE writer - spi cs line keep low when the bit is set.

MASTER_CK_SEL_R

Field MASTER_CK_SEL reader - In the master mode spi cs line is enable as spi clk it is combined with spi_cs0_dis spi_cs1_dis spi_cs2_dis.

MASTER_CK_SEL_W

Field MASTER_CK_SEL writer - In the master mode spi cs line is enable as spi clk it is combined with spi_cs0_dis spi_cs1_dis spi_cs2_dis.

MASTER_CS_POL_R

Field MASTER_CS_POL reader - In the master mode the bits are the polarity of spi cs line the value is equivalent to spi_cs ^ spi_master_cs_pol.

MASTER_CS_POL_W

Field MASTER_CS_POL writer - In the master mode the bits are the polarity of spi cs line the value is equivalent to spi_cs ^ spi_master_cs_pol.

R

Register PIN reader

W

Register PIN writer