Module esp32::spi0::slave

Structs

SLAVE_SPEC

You can read this register and get slave::R. You can reset, write, write_with_zero this register using slave::W. You can also modify this register. See API.

Type Aliases

CS_I_MODE_R

Field CS_I_MODE reader - In the slave mode this bits used to synchronize the input spi cs signal and eliminate spi cs jitter.

CS_I_MODE_W

Field CS_I_MODE writer - In the slave mode this bits used to synchronize the input spi cs signal and eliminate spi cs jitter.

INT_EN_R

Field INT_EN reader - Interrupt enable bits for the below 5 sources

INT_EN_W

Field INT_EN writer - Interrupt enable bits for the below 5 sources

MODE_R

Field MODE reader - 1: slave mode 0: master mode.

MODE_W

Field MODE writer - 1: slave mode 0: master mode.

R

Register SLAVE reader

SLV_CMD_DEFINE_R

Field SLV_CMD_DEFINE reader - 1: slave mode commands are defined in SPI_SLAVE3. 0: slave mode commands are fixed as: 1: write-status 2: write-buffer and 3: read-buffer.

SLV_CMD_DEFINE_W

Field SLV_CMD_DEFINE writer - 1: slave mode commands are defined in SPI_SLAVE3. 0: slave mode commands are fixed as: 1: write-status 2: write-buffer and 3: read-buffer.

SLV_LAST_COMMAND_R

Field SLV_LAST_COMMAND reader - In the slave mode it is the value of command.

SLV_LAST_STATE_R

Field SLV_LAST_STATE reader - In the slave mode it is the state of spi state machine.

SLV_RD_BUF_DONE_R

Field SLV_RD_BUF_DONE reader - The interrupt raw bit for the completion of read-buffer operation in the slave mode.

SLV_RD_BUF_DONE_W

Field SLV_RD_BUF_DONE writer - The interrupt raw bit for the completion of read-buffer operation in the slave mode.

SLV_RD_STA_DONE_R

Field SLV_RD_STA_DONE reader - The interrupt raw bit for the completion of read-status operation in the slave mode.

SLV_RD_STA_DONE_W

Field SLV_RD_STA_DONE writer - The interrupt raw bit for the completion of read-status operation in the slave mode.

SLV_WR_BUF_DONE_R

Field SLV_WR_BUF_DONE reader - The interrupt raw bit for the completion of write-buffer operation in the slave mode.

SLV_WR_BUF_DONE_W

Field SLV_WR_BUF_DONE writer - The interrupt raw bit for the completion of write-buffer operation in the slave mode.

SLV_WR_RD_BUF_EN_R

Field SLV_WR_RD_BUF_EN reader - write and read buffer enable in the slave mode

SLV_WR_RD_BUF_EN_W

Field SLV_WR_RD_BUF_EN writer - write and read buffer enable in the slave mode

SLV_WR_RD_STA_EN_R

Field SLV_WR_RD_STA_EN reader - write and read status enable in the slave mode

SLV_WR_RD_STA_EN_W

Field SLV_WR_RD_STA_EN writer - write and read status enable in the slave mode

SLV_WR_STA_DONE_R

Field SLV_WR_STA_DONE reader - The interrupt raw bit for the completion of write-status operation in the slave mode.

SLV_WR_STA_DONE_W

Field SLV_WR_STA_DONE writer - The interrupt raw bit for the completion of write-status operation in the slave mode.

SYNC_RESET_R

Field SYNC_RESET reader - Software reset enable, reset the spi clock line cs line and data lines.

SYNC_RESET_W

Field SYNC_RESET writer - Software reset enable, reset the spi clock line cs line and data lines.

TRANS_CNT_R

Field TRANS_CNT reader - The operations counter in both the master mode and the slave mode. 4: read-status

TRANS_DONE_R

Field TRANS_DONE reader - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode.

TRANS_DONE_W

Field TRANS_DONE writer - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode.

W

Register SLAVE writer