Type Alias esp32::spi0::slave::R

pub type R = R<SLAVE_SPEC>;

Register SLAVE reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn slv_rd_buf_done(&self) -> SLV_RD_BUF_DONE_R

Bit 0 - The interrupt raw bit for the completion of read-buffer operation in the slave mode.

pub fn slv_wr_buf_done(&self) -> SLV_WR_BUF_DONE_R

Bit 1 - The interrupt raw bit for the completion of write-buffer operation in the slave mode.

pub fn slv_rd_sta_done(&self) -> SLV_RD_STA_DONE_R

Bit 2 - The interrupt raw bit for the completion of read-status operation in the slave mode.

pub fn slv_wr_sta_done(&self) -> SLV_WR_STA_DONE_R

Bit 3 - The interrupt raw bit for the completion of write-status operation in the slave mode.

pub fn trans_done(&self) -> TRANS_DONE_R

Bit 4 - The interrupt raw bit for the completion of any operation in both the master mode and the slave mode.

pub fn int_en(&self) -> INT_EN_R

Bits 5:9 - Interrupt enable bits for the below 5 sources

pub fn cs_i_mode(&self) -> CS_I_MODE_R

Bits 10:11 - In the slave mode this bits used to synchronize the input spi cs signal and eliminate spi cs jitter.

pub fn slv_last_command(&self) -> SLV_LAST_COMMAND_R

Bits 17:19 - In the slave mode it is the value of command.

pub fn slv_last_state(&self) -> SLV_LAST_STATE_R

Bits 20:22 - In the slave mode it is the state of spi state machine.

pub fn trans_cnt(&self) -> TRANS_CNT_R

Bits 23:26 - The operations counter in both the master mode and the slave mode. 4: read-status

pub fn slv_cmd_define(&self) -> SLV_CMD_DEFINE_R

Bit 27 - 1: slave mode commands are defined in SPI_SLAVE3. 0: slave mode commands are fixed as: 1: write-status 2: write-buffer and 3: read-buffer.

pub fn slv_wr_rd_sta_en(&self) -> SLV_WR_RD_STA_EN_R

Bit 28 - write and read status enable in the slave mode

pub fn slv_wr_rd_buf_en(&self) -> SLV_WR_RD_BUF_EN_R

Bit 29 - write and read buffer enable in the slave mode

pub fn mode(&self) -> MODE_R

Bit 30 - 1: slave mode 0: master mode.

pub fn sync_reset(&self) -> SYNC_RESET_R

Bit 31 - Software reset enable, reset the spi clock line cs line and data lines.