Module esp32::uart0::int_raw

Structs

INT_RAW_SPEC

You can read this register and get int_raw::R. See API.

Type Aliases

AT_CMD_CHAR_DET_INT_RAW_R

Field AT_CMD_CHAR_DET_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the configured at_cmd chars.

BRK_DET_INT_RAW_R

Field BRK_DET_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the 0 after the stop bit.

CTS_CHG_INT_RAW_R

Field CTS_CHG_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the edge change of ctsn signal.

DSR_CHG_INT_RAW_R

Field DSR_CHG_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the edge change of dsrn signal.

FRM_ERR_INT_RAW_R

Field FRM_ERR_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects data’s frame error .

GLITCH_DET_INT_RAW_R

Field GLITCH_DET_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the start bit.

PARITY_ERR_INT_RAW_R

Field PARITY_ERR_INT_RAW reader - This interrupt raw bit turns to high level when receiver detects the parity error of data.

R

Register INT_RAW reader

RS485_CLASH_INT_RAW_R

Field RS485_CLASH_INT_RAW reader - This interrupt raw bit turns to high level when rs485 detects the clash between transmitter and receiver.

RS485_FRM_ERR_INT_RAW_R

Field RS485_FRM_ERR_INT_RAW reader - This interrupt raw bit turns to high level when rs485 detects the data frame error.

RS485_PARITY_ERR_INT_RAW_R

Field RS485_PARITY_ERR_INT_RAW reader - This interrupt raw bit turns to high level when rs485 detects the parity error.

RXFIFO_FULL_INT_RAW_R

Field RXFIFO_FULL_INT_RAW reader - This interrupt raw bit turns to high level when receiver receives more data than (rx_flow_thrhd_h3 rx_flow_thrhd).

RXFIFO_OVF_INT_RAW_R

Field RXFIFO_OVF_INT_RAW reader - This interrupt raw bit turns to high level when receiver receives more data than the fifo can store.

RXFIFO_TOUT_INT_RAW_R

Field RXFIFO_TOUT_INT_RAW reader - This interrupt raw bit turns to high level when receiver takes more time than rx_tout_thrhd to receive a byte.

SW_XOFF_INT_RAW_R

Field SW_XOFF_INT_RAW reader - This interrupt raw bit turns to high level when receiver receives xon char with uart_sw_flow_con_en is set to 1.

SW_XON_INT_RAW_R

Field SW_XON_INT_RAW reader - This interrupt raw bit turns to high level when receiver receives xoff char with uart_sw_flow_con_en is set to 1.

TXFIFO_EMPTY_INT_RAW_R

Field TXFIFO_EMPTY_INT_RAW reader - This interrupt raw bit turns to high level when the amount of data in transmitter’s fifo is less than ((tx_mem_cnttxfifo_cnt) .

TX_BRK_DONE_INT_RAW_R

Field TX_BRK_DONE_INT_RAW reader - This interrupt raw bit turns to high level when transmitter completes sendding 0 after all the datas in transmitter’s fifo are send.

TX_BRK_IDLE_DONE_INT_RAW_R

Field TX_BRK_IDLE_DONE_INT_RAW reader - This interrupt raw bit turns to high level when transmitter has kept the shortest duration after the last data has been send.

TX_DONE_INT_RAW_R

Field TX_DONE_INT_RAW reader - This interrupt raw bit turns to high level when transmitter has send all the data in fifo.