Type Alias esp32::uart0::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rxfifo_full_int_raw(&self) -> RXFIFO_FULL_INT_RAW_R

Bit 0 - This interrupt raw bit turns to high level when receiver receives more data than (rx_flow_thrhd_h3 rx_flow_thrhd).

pub fn txfifo_empty_int_raw(&self) -> TXFIFO_EMPTY_INT_RAW_R

Bit 1 - This interrupt raw bit turns to high level when the amount of data in transmitter’s fifo is less than ((tx_mem_cnttxfifo_cnt) .

pub fn parity_err_int_raw(&self) -> PARITY_ERR_INT_RAW_R

Bit 2 - This interrupt raw bit turns to high level when receiver detects the parity error of data.

pub fn frm_err_int_raw(&self) -> FRM_ERR_INT_RAW_R

Bit 3 - This interrupt raw bit turns to high level when receiver detects data’s frame error .

pub fn rxfifo_ovf_int_raw(&self) -> RXFIFO_OVF_INT_RAW_R

Bit 4 - This interrupt raw bit turns to high level when receiver receives more data than the fifo can store.

pub fn dsr_chg_int_raw(&self) -> DSR_CHG_INT_RAW_R

Bit 5 - This interrupt raw bit turns to high level when receiver detects the edge change of dsrn signal.

pub fn cts_chg_int_raw(&self) -> CTS_CHG_INT_RAW_R

Bit 6 - This interrupt raw bit turns to high level when receiver detects the edge change of ctsn signal.

pub fn brk_det_int_raw(&self) -> BRK_DET_INT_RAW_R

Bit 7 - This interrupt raw bit turns to high level when receiver detects the 0 after the stop bit.

pub fn rxfifo_tout_int_raw(&self) -> RXFIFO_TOUT_INT_RAW_R

Bit 8 - This interrupt raw bit turns to high level when receiver takes more time than rx_tout_thrhd to receive a byte.

pub fn sw_xon_int_raw(&self) -> SW_XON_INT_RAW_R

Bit 9 - This interrupt raw bit turns to high level when receiver receives xoff char with uart_sw_flow_con_en is set to 1.

pub fn sw_xoff_int_raw(&self) -> SW_XOFF_INT_RAW_R

Bit 10 - This interrupt raw bit turns to high level when receiver receives xon char with uart_sw_flow_con_en is set to 1.

pub fn glitch_det_int_raw(&self) -> GLITCH_DET_INT_RAW_R

Bit 11 - This interrupt raw bit turns to high level when receiver detects the start bit.

pub fn tx_brk_done_int_raw(&self) -> TX_BRK_DONE_INT_RAW_R

Bit 12 - This interrupt raw bit turns to high level when transmitter completes sendding 0 after all the datas in transmitter’s fifo are send.

pub fn tx_brk_idle_done_int_raw(&self) -> TX_BRK_IDLE_DONE_INT_RAW_R

Bit 13 - This interrupt raw bit turns to high level when transmitter has kept the shortest duration after the last data has been send.

pub fn tx_done_int_raw(&self) -> TX_DONE_INT_RAW_R

Bit 14 - This interrupt raw bit turns to high level when transmitter has send all the data in fifo.

pub fn rs485_parity_err_int_raw(&self) -> RS485_PARITY_ERR_INT_RAW_R

Bit 15 - This interrupt raw bit turns to high level when rs485 detects the parity error.

pub fn rs485_frm_err_int_raw(&self) -> RS485_FRM_ERR_INT_RAW_R

Bit 16 - This interrupt raw bit turns to high level when rs485 detects the data frame error.

pub fn rs485_clash_int_raw(&self) -> RS485_CLASH_INT_RAW_R

Bit 17 - This interrupt raw bit turns to high level when rs485 detects the clash between transmitter and receiver.

pub fn at_cmd_char_det_int_raw(&self) -> AT_CMD_CHAR_DET_INT_RAW_R

Bit 18 - This interrupt raw bit turns to high level when receiver detects the configured at_cmd chars.