Module esp32::uart0::mem_conf

Structs

MEM_CONF_SPEC

You can read this register and get mem_conf::R. You can reset, write, write_with_zero this register using mem_conf::W. You can also modify this register. See API.

Type Aliases

MEM_PD_R

Field MEM_PD reader - Set this bit to power down mem.when reg_mem_pd registers in the 3 uarts are all set to 1 mem will enter low power mode.

MEM_PD_W

Field MEM_PD writer - Set this bit to power down mem.when reg_mem_pd registers in the 3 uarts are all set to 1 mem will enter low power mode.

R

Register MEM_CONF reader

RX_FLOW_THRHD_H3_R

Field RX_FLOW_THRHD_H3 reader - Refer to the rx_flow_thrhd’s description.

RX_FLOW_THRHD_H3_W

Field RX_FLOW_THRHD_H3 writer - Refer to the rx_flow_thrhd’s description.

RX_MEM_FULL_THRHD_R

Field RX_MEM_FULL_THRHD reader - Refer to the rxfifo_full_thrhd’s description.

RX_MEM_FULL_THRHD_W

Field RX_MEM_FULL_THRHD writer - Refer to the rxfifo_full_thrhd’s description.

RX_SIZE_R

Field RX_SIZE reader - This register is used to configure the amount of mem allocated to receiver’s fifo. the default byte num is 128.

RX_SIZE_W

Field RX_SIZE writer - This register is used to configure the amount of mem allocated to receiver’s fifo. the default byte num is 128.

RX_TOUT_THRHD_H3_R

Field RX_TOUT_THRHD_H3 reader - Refer to the rx_tout_thrhd’s description.

RX_TOUT_THRHD_H3_W

Field RX_TOUT_THRHD_H3 writer - Refer to the rx_tout_thrhd’s description.

TX_MEM_EMPTY_THRHD_R

Field TX_MEM_EMPTY_THRHD reader - Refer to txfifo_empty_thrhd’s description.

TX_MEM_EMPTY_THRHD_W

Field TX_MEM_EMPTY_THRHD writer - Refer to txfifo_empty_thrhd’s description.

TX_SIZE_R

Field TX_SIZE reader - This register is used to configure the amount of mem allocated to transmitter’s fifo.the default byte num is 128.

TX_SIZE_W

Field TX_SIZE writer - This register is used to configure the amount of mem allocated to transmitter’s fifo.the default byte num is 128.

W

Register MEM_CONF writer

XOFF_THRESHOLD_H2_R

Field XOFF_THRESHOLD_H2 reader - Refer to the uart_xoff_threshold’s description.

XOFF_THRESHOLD_H2_W

Field XOFF_THRESHOLD_H2 writer - Refer to the uart_xoff_threshold’s description.

XON_THRESHOLD_H2_R

Field XON_THRESHOLD_H2 reader - Refer to the uart_xon_threshold’s description.

XON_THRESHOLD_H2_W

Field XON_THRESHOLD_H2 writer - Refer to the uart_xon_threshold’s description.