Type Alias esp32::uart0::mem_conf::W

pub type W = W<MEM_CONF_SPEC>;

Register MEM_CONF writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn mem_pd(&mut self) -> MEM_PD_W<'_, MEM_CONF_SPEC>

Bit 0 - Set this bit to power down mem.when reg_mem_pd registers in the 3 uarts are all set to 1 mem will enter low power mode.

pub fn rx_size(&mut self) -> RX_SIZE_W<'_, MEM_CONF_SPEC>

Bits 3:6 - This register is used to configure the amount of mem allocated to receiver’s fifo. the default byte num is 128.

pub fn tx_size(&mut self) -> TX_SIZE_W<'_, MEM_CONF_SPEC>

Bits 7:10 - This register is used to configure the amount of mem allocated to transmitter’s fifo.the default byte num is 128.

pub fn rx_flow_thrhd_h3(&mut self) -> RX_FLOW_THRHD_H3_W<'_, MEM_CONF_SPEC>

Bits 15:17 - Refer to the rx_flow_thrhd’s description.

pub fn rx_tout_thrhd_h3(&mut self) -> RX_TOUT_THRHD_H3_W<'_, MEM_CONF_SPEC>

Bits 18:20 - Refer to the rx_tout_thrhd’s description.

pub fn xon_threshold_h2(&mut self) -> XON_THRESHOLD_H2_W<'_, MEM_CONF_SPEC>

Bits 21:22 - Refer to the uart_xon_threshold’s description.

pub fn xoff_threshold_h2(&mut self) -> XOFF_THRESHOLD_H2_W<'_, MEM_CONF_SPEC>

Bits 23:24 - Refer to the uart_xoff_threshold’s description.

pub fn rx_mem_full_thrhd(&mut self) -> RX_MEM_FULL_THRHD_W<'_, MEM_CONF_SPEC>

Bits 25:27 - Refer to the rxfifo_full_thrhd’s description.

pub fn tx_mem_empty_thrhd(&mut self) -> TX_MEM_EMPTY_THRHD_W<'_, MEM_CONF_SPEC>

Bits 28:30 - Refer to txfifo_empty_thrhd’s description.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual