Module esp32::uart0::rs485_conf

Structs

RS485_CONF_SPEC

You can read this register and get rs485_conf::R. You can reset, write, write_with_zero this register using rs485_conf::W. You can also modify this register. See API.

Type Aliases

DL0_EN_R

Field DL0_EN reader - Set this bit to delay the stop bit by 1 bit.

DL0_EN_W

Field DL0_EN writer - Set this bit to delay the stop bit by 1 bit.

DL1_EN_R

Field DL1_EN reader - Set this bit to delay the stop bit by 1 bit.

DL1_EN_W

Field DL1_EN writer - Set this bit to delay the stop bit by 1 bit.

R

Register RS485_CONF reader

RS485RXBY_TX_EN_R

Field RS485RXBY_TX_EN reader - 1: enable rs485’s transmitter to send data when rs485’s receiver is busy. 0:rs485’s transmitter should not send data when its receiver is busy.

RS485RXBY_TX_EN_W

Field RS485RXBY_TX_EN writer - 1: enable rs485’s transmitter to send data when rs485’s receiver is busy. 0:rs485’s transmitter should not send data when its receiver is busy.

RS485TX_RX_EN_R

Field RS485TX_RX_EN reader - Set this bit to enable loopback transmitter’s output data signal to receiver’s input data signal.

RS485TX_RX_EN_W

Field RS485TX_RX_EN writer - Set this bit to enable loopback transmitter’s output data signal to receiver’s input data signal.

RS485_EN_R

Field RS485_EN reader - Set this bit to choose rs485 mode.

RS485_EN_W

Field RS485_EN writer - Set this bit to choose rs485 mode.

RS485_RX_DLY_NUM_R

Field RS485_RX_DLY_NUM reader - This register is used to delay the receiver’s internal data signal.

RS485_RX_DLY_NUM_W

Field RS485_RX_DLY_NUM writer - This register is used to delay the receiver’s internal data signal.

RS485_TX_DLY_NUM_R

Field RS485_TX_DLY_NUM reader - This register is used to delay the transmitter’s internal data signal.

RS485_TX_DLY_NUM_W

Field RS485_TX_DLY_NUM writer - This register is used to delay the transmitter’s internal data signal.

W

Register RS485_CONF writer