Module esp32::uhci0::conf0

Structs

CONF0_SPEC

You can read this register and get conf0::R. You can reset, write, write_with_zero this register using conf0::W. You can also modify this register. See API.

Type Aliases

AHBM_FIFO_RST_R

Field AHBM_FIFO_RST reader - Set this bit to reset dma ahb fifo.

AHBM_FIFO_RST_W

Field AHBM_FIFO_RST writer - Set this bit to reset dma ahb fifo.

AHBM_RST_R

Field AHBM_RST reader - Set this bit to reset dma ahb interface.

AHBM_RST_W

Field AHBM_RST writer - Set this bit to reset dma ahb interface.

CLK_EN_R

Field CLK_EN reader - Set this bit to enable clock-gating for read or write registers.

CLK_EN_W

Field CLK_EN writer - Set this bit to enable clock-gating for read or write registers.

CRC_REC_EN_R

Field CRC_REC_EN reader - Set this bit to enable receiver’’s ability of crc calculation when crc_en bit in head packet is 1 then there will be crc bytes after data_frame

CRC_REC_EN_W

Field CRC_REC_EN writer - Set this bit to enable receiver’’s ability of crc calculation when crc_en bit in head packet is 1 then there will be crc bytes after data_frame

ENCODE_CRC_EN_R

Field ENCODE_CRC_EN reader - Set this bit to enable crc calculation for data frame when bit6 in the head packet is 1.

ENCODE_CRC_EN_W

Field ENCODE_CRC_EN writer - Set this bit to enable crc calculation for data frame when bit6 in the head packet is 1.

HEAD_EN_R

Field HEAD_EN reader - Set this bit to enable to use head packet before the data frame.

HEAD_EN_W

Field HEAD_EN writer - Set this bit to enable to use head packet before the data frame.

INDSCR_BURST_EN_R

Field INDSCR_BURST_EN reader - Set this bit to enable DMA out links to use burst mode.

INDSCR_BURST_EN_W

Field INDSCR_BURST_EN writer - Set this bit to enable DMA out links to use burst mode.

IN_LOOP_TEST_R

Field IN_LOOP_TEST reader - Set this bit to enable loop test for in links.

IN_LOOP_TEST_W

Field IN_LOOP_TEST writer - Set this bit to enable loop test for in links.

IN_RST_R

Field IN_RST reader - Set this bit to reset in link operations.

IN_RST_W

Field IN_RST writer - Set this bit to reset in link operations.

LEN_EOF_EN_R

Field LEN_EOF_EN reader - Set this bit to enable to use packet_len in packet head when the received data is equal to packet_len this means the end of a data frame.

LEN_EOF_EN_W

Field LEN_EOF_EN writer - Set this bit to enable to use packet_len in packet head when the received data is equal to packet_len this means the end of a data frame.

MEM_TRANS_EN_R

Field MEM_TRANS_EN reader -

MEM_TRANS_EN_W

Field MEM_TRANS_EN writer -

OUTDSCR_BURST_EN_R

Field OUTDSCR_BURST_EN reader - Set this bit to enable DMA in links to use burst mode.

OUTDSCR_BURST_EN_W

Field OUTDSCR_BURST_EN writer - Set this bit to enable DMA in links to use burst mode.

OUT_AUTO_WRBACK_R

Field OUT_AUTO_WRBACK reader - when in link’s length is 0 go on to use the next in link automatically.

OUT_AUTO_WRBACK_W

Field OUT_AUTO_WRBACK writer - when in link’s length is 0 go on to use the next in link automatically.

OUT_DATA_BURST_EN_R

Field OUT_DATA_BURST_EN reader - Set this bit to enable DMA burst MODE

OUT_DATA_BURST_EN_W

Field OUT_DATA_BURST_EN writer - Set this bit to enable DMA burst MODE

OUT_EOF_MODE_R

Field OUT_EOF_MODE reader - Set this bit to produce eof after DMA pops all data clear this bit to produce eof after DMA pushes all data

OUT_EOF_MODE_W

Field OUT_EOF_MODE writer - Set this bit to produce eof after DMA pops all data clear this bit to produce eof after DMA pushes all data

OUT_LOOP_TEST_R

Field OUT_LOOP_TEST reader - Set this bit to enable loop test for out links.

OUT_LOOP_TEST_W

Field OUT_LOOP_TEST writer - Set this bit to enable loop test for out links.

OUT_NO_RESTART_CLR_R

Field OUT_NO_RESTART_CLR reader - don’t use

OUT_NO_RESTART_CLR_W

Field OUT_NO_RESTART_CLR writer - don’t use

OUT_RST_R

Field OUT_RST reader - Set this bit to reset out link operations.

OUT_RST_W

Field OUT_RST writer - Set this bit to reset out link operations.

R

Register CONF0 reader

SEPER_EN_R

Field SEPER_EN reader - Set this bit to use special char to separate the data frame.

SEPER_EN_W

Field SEPER_EN writer - Set this bit to use special char to separate the data frame.

UART0_CE_R

Field UART0_CE reader - Set this bit to use UART to transmit or receive data.

UART0_CE_W

Field UART0_CE writer - Set this bit to use UART to transmit or receive data.

UART1_CE_R

Field UART1_CE reader - Set this bit to use UART1 to transmit or receive data.

UART1_CE_W

Field UART1_CE writer - Set this bit to use UART1 to transmit or receive data.

UART2_CE_R

Field UART2_CE reader - Set this bit to use UART2 to transmit or receive data.

UART2_CE_W

Field UART2_CE writer - Set this bit to use UART2 to transmit or receive data.

UART_IDLE_EOF_EN_R

Field UART_IDLE_EOF_EN reader - Set this bit to enable to use idle time when the idle time after data frame is satisfied this means the end of a data frame.

UART_IDLE_EOF_EN_W

Field UART_IDLE_EOF_EN writer - Set this bit to enable to use idle time when the idle time after data frame is satisfied this means the end of a data frame.

UART_RX_BRK_EOF_EN_R

Field UART_RX_BRK_EOF_EN reader - Set this bit to enable to use brk char as the end of a data frame.

UART_RX_BRK_EOF_EN_W

Field UART_RX_BRK_EOF_EN writer - Set this bit to enable to use brk char as the end of a data frame.

W

Register CONF0 writer