Type Alias esp32::uhci0::conf0::W

pub type W = W<CONF0_SPEC>;

Register CONF0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn in_rst(&mut self) -> IN_RST_W<'_, CONF0_SPEC>

Bit 0 - Set this bit to reset in link operations.

pub fn out_rst(&mut self) -> OUT_RST_W<'_, CONF0_SPEC>

Bit 1 - Set this bit to reset out link operations.

pub fn ahbm_fifo_rst(&mut self) -> AHBM_FIFO_RST_W<'_, CONF0_SPEC>

Bit 2 - Set this bit to reset dma ahb fifo.

pub fn ahbm_rst(&mut self) -> AHBM_RST_W<'_, CONF0_SPEC>

Bit 3 - Set this bit to reset dma ahb interface.

pub fn in_loop_test(&mut self) -> IN_LOOP_TEST_W<'_, CONF0_SPEC>

Bit 4 - Set this bit to enable loop test for in links.

pub fn out_loop_test(&mut self) -> OUT_LOOP_TEST_W<'_, CONF0_SPEC>

Bit 5 - Set this bit to enable loop test for out links.

pub fn out_auto_wrback(&mut self) -> OUT_AUTO_WRBACK_W<'_, CONF0_SPEC>

Bit 6 - when in link’s length is 0 go on to use the next in link automatically.

pub fn out_no_restart_clr(&mut self) -> OUT_NO_RESTART_CLR_W<'_, CONF0_SPEC>

Bit 7 - don’t use

pub fn out_eof_mode(&mut self) -> OUT_EOF_MODE_W<'_, CONF0_SPEC>

Bit 8 - Set this bit to produce eof after DMA pops all data clear this bit to produce eof after DMA pushes all data

pub fn uart0_ce(&mut self) -> UART0_CE_W<'_, CONF0_SPEC>

Bit 9 - Set this bit to use UART to transmit or receive data.

pub fn uart1_ce(&mut self) -> UART1_CE_W<'_, CONF0_SPEC>

Bit 10 - Set this bit to use UART1 to transmit or receive data.

pub fn uart2_ce(&mut self) -> UART2_CE_W<'_, CONF0_SPEC>

Bit 11 - Set this bit to use UART2 to transmit or receive data.

pub fn outdscr_burst_en(&mut self) -> OUTDSCR_BURST_EN_W<'_, CONF0_SPEC>

Bit 12 - Set this bit to enable DMA in links to use burst mode.

pub fn indscr_burst_en(&mut self) -> INDSCR_BURST_EN_W<'_, CONF0_SPEC>

Bit 13 - Set this bit to enable DMA out links to use burst mode.

pub fn out_data_burst_en(&mut self) -> OUT_DATA_BURST_EN_W<'_, CONF0_SPEC>

Bit 14 - Set this bit to enable DMA burst MODE

pub fn mem_trans_en(&mut self) -> MEM_TRANS_EN_W<'_, CONF0_SPEC>

Bit 15

pub fn seper_en(&mut self) -> SEPER_EN_W<'_, CONF0_SPEC>

Bit 16 - Set this bit to use special char to separate the data frame.

pub fn head_en(&mut self) -> HEAD_EN_W<'_, CONF0_SPEC>

Bit 17 - Set this bit to enable to use head packet before the data frame.

pub fn crc_rec_en(&mut self) -> CRC_REC_EN_W<'_, CONF0_SPEC>

Bit 18 - Set this bit to enable receiver’’s ability of crc calculation when crc_en bit in head packet is 1 then there will be crc bytes after data_frame

pub fn uart_idle_eof_en(&mut self) -> UART_IDLE_EOF_EN_W<'_, CONF0_SPEC>

Bit 19 - Set this bit to enable to use idle time when the idle time after data frame is satisfied this means the end of a data frame.

pub fn len_eof_en(&mut self) -> LEN_EOF_EN_W<'_, CONF0_SPEC>

Bit 20 - Set this bit to enable to use packet_len in packet head when the received data is equal to packet_len this means the end of a data frame.

pub fn encode_crc_en(&mut self) -> ENCODE_CRC_EN_W<'_, CONF0_SPEC>

Bit 21 - Set this bit to enable crc calculation for data frame when bit6 in the head packet is 1.

pub fn clk_en(&mut self) -> CLK_EN_W<'_, CONF0_SPEC>

Bit 22 - Set this bit to enable clock-gating for read or write registers.

pub fn uart_rx_brk_eof_en(&mut self) -> UART_RX_BRK_EOF_EN_W<'_, CONF0_SPEC>

Bit 23 - Set this bit to enable to use brk char as the end of a data frame.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual