Module esp32::uhci0::conf1

Structs

CONF1_SPEC

You can read this register and get conf1::R. You can reset, write, write_with_zero this register using conf1::W. You can also modify this register. See API.

Type Aliases

CHECK_OWNER_R

Field CHECK_OWNER reader - Set this bit to check the owner bit in link descriptor.

CHECK_OWNER_W

Field CHECK_OWNER writer - Set this bit to check the owner bit in link descriptor.

CHECK_SEQ_EN_R

Field CHECK_SEQ_EN reader - Set this bit to enable decoder to check seq num in packet header.

CHECK_SEQ_EN_W

Field CHECK_SEQ_EN writer - Set this bit to enable decoder to check seq num in packet header.

CHECK_SUM_EN_R

Field CHECK_SUM_EN reader - Set this bit to enable decoder to check check_sum in packet header.

CHECK_SUM_EN_W

Field CHECK_SUM_EN writer - Set this bit to enable decoder to check check_sum in packet header.

CRC_DISABLE_R

Field CRC_DISABLE reader - Set this bit to disable crc calculation.

CRC_DISABLE_W

Field CRC_DISABLE writer - Set this bit to disable crc calculation.

DMA_INFIFO_FULL_THRS_R

Field DMA_INFIFO_FULL_THRS reader - when data amount in link descriptor’s fifo is more than this register value it will produce uhci_dma_infifo_full_wm_int interrupt.

DMA_INFIFO_FULL_THRS_W

Field DMA_INFIFO_FULL_THRS writer - when data amount in link descriptor’s fifo is more than this register value it will produce uhci_dma_infifo_full_wm_int interrupt.

R

Register CONF1 reader

SAVE_HEAD_R

Field SAVE_HEAD reader - Set this bit to save packet header .

SAVE_HEAD_W

Field SAVE_HEAD writer - Set this bit to save packet header .

SW_START_R

Field SW_START reader - Set this bit to start inserting the packet header.

SW_START_W

Field SW_START writer - Set this bit to start inserting the packet header.

TX_ACK_NUM_RE_R

Field TX_ACK_NUM_RE reader - Set this bit to enable hardware replace ack num in packet header automatically.

TX_ACK_NUM_RE_W

Field TX_ACK_NUM_RE writer - Set this bit to enable hardware replace ack num in packet header automatically.

TX_CHECK_SUM_RE_R

Field TX_CHECK_SUM_RE reader - Set this bit to enable hardware replace check_sum in packet header automatically.

TX_CHECK_SUM_RE_W

Field TX_CHECK_SUM_RE writer - Set this bit to enable hardware replace check_sum in packet header automatically.

W

Register CONF1 writer

WAIT_SW_START_R

Field WAIT_SW_START reader - Set this bit to enable software way to add packet header.

WAIT_SW_START_W

Field WAIT_SW_START writer - Set this bit to enable software way to add packet header.