Type Alias esp32::uhci0::dma_in_link::R

pub type R = R<DMA_IN_LINK_SPEC>;

Register DMA_IN_LINK reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn inlink_addr(&self) -> INLINK_ADDR_R

Bits 0:19 - This register stores the least 20 bits of the first in link descriptor’s address.

pub fn inlink_auto_ret(&self) -> INLINK_AUTO_RET_R

Bit 20 - 1:when a packet is wrong in link descriptor returns to the descriptor which is lately used.

pub fn inlink_stop(&self) -> INLINK_STOP_R

Bit 28 - Set this bit to stop dealing with the in link descriptors.

pub fn inlink_start(&self) -> INLINK_START_R

Bit 29 - Set this bit to start dealing with the in link descriptors.

pub fn inlink_restart(&self) -> INLINK_RESTART_R

Bit 30 - Set this bit to mount on new in link descriptors

pub fn inlink_park(&self) -> INLINK_PARK_R

Bit 31 - 1:the in link descriptor’s fsm is in idle state. 0:the in link descriptor’s fsm is working